Standard Arguments

The documentation to this function lists all the arguments in random.cdisc.data that are used repeatedly in dataset creation.

Arguments

seed

(numeric)
Seed to use for reproducible random number generation.

na_percentage

(proportion)
Default percentage of values to be replaced by NA.

na_vars

(list)
A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements:

seed (numeric)
The seed to be used for this element - can be NA.

percentage (proportion)
Percentage of elements to be replaced with NA. If NA, na_percentage is used as a default.

adsl

(data.frame)
Subject-Level Analysis Dataset (ADSL).

lookup

(data.frame)
Additional parameters.

lookup_aag

(data.frame)
Additional metadata parameters.

param

(character vector)
Parameter values.

paramcd

(character vector)
Parameter code values.

paramu

(character vector)
Parameter unit values.

visit_format

(character)
Type of visit. Options are "WEEK" and "CYCLE".

n_assessments

(integer)
Number of weeks or cycles.

n_days

(integer)
Number of days in each cycle (only used if visit_format is "CYCLE").

aval_mean

(numeric vector)
Mean values corresponding to each parameter.