Helper Functions for Constructing ADQLQC

Internal functions used by radqlqc.

Usage

get_qs_data(
  adsl,
  visit_format = "CYCLE",
  n_assessments = 5L,
  n_days = 1L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(QSORRES = c(1234, 0.2), QSSTRESC = c(1234, 0.2))
)

get_random_dates_between(from, to, visit_id)

prep_adqlqc(df)

calc_scales(adqlqc1)

derv_chgcat1(dataset)

comp_derv(dataset, percent, number)

Arguments

adsl

(data.frame)
Subject-Level Analysis Dataset (ADSL).

visit_format

(character)
Type of visit. Options are "WEEK" and "CYCLE".

n_assessments

(integer)
Number of weeks or cycles.

n_days

(integer)
Number of days in each cycle (only used if visit_format is "CYCLE").

lookup

(data.frame)
Additional parameters.

seed

(numeric)
Seed to use for reproducible random number generation.

na_percentage

(proportion)
Default percentage of values to be replaced by NA.

na_vars

(list)
A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements:

seed (numeric)
The seed to be used for this element - can be NA.

percentage (proportion)
Percentage of elements to be replaced with NA. If NA, na_percentage is used as a default.

from

(datetime vector)
Start date/times.

to

(datetime vector)
End date/times.

visit_id

(vector)
Visit identifiers.

df

(data.frame)
SDTM QS dataset.

adqlqc1

(data.frame)
Prepared data generated from the prep_adqlqc() function.

dataset

(data.frame)
Dataset.

percent

(numeric)
Completion - Completed at least y percent of questions, 1 record per visit

number

(numeric)
Completion - Completed at least x question(s), 1 record per visit

Value

a dataframe with SDTM questionnaire data

Data frame with new randomly generated dates variable.

data.frame

data.frame

data.frame

data.frame

Functions

get_qs_data(): Questionnaires EORTC QLQ-C30 V3.0 SDTM (QS)

Function for generating random Questionnaires SDTM domain

get_random_dates_between(): Function for generating random dates between 2 dates

prep_adqlqc(): Prepare ADaM ADQLQC data, adding PARAMCD to SDTM QS data

calc_scales(): Scale calculation for ADQLQC data

derv_chgcat1(): Calculate Change from Baseline Category 1

comp_derv(): Completion/Compliance Data Calculation

Examples

adsl <- radsl(N = 10, study_duration = 2, seed = 1)
adqlqc <- radqlqc(adsl, seed = 1, percent = 80, number = 2)

qs <- random.cdisc.data:::get_qs_data(adsl, n_assessments = 5L, seed = 1, na_percentage = 0.1)
qs
#> # A tibble: 1,771 × 17
#>    STUDYID USUBJID  QSSEQ QSTESTCD QSTEST QSCAT QSSCAT QSORRES QSORRESU QSSTRESC
#>    <chr>   <chr>    <int> <chr>    <chr>  <chr> <chr>  <chr>   <chr>    <chr>   
#>  1 AB12345 AB12345…     1 EOR0101  EOR01… EORT… NA     NOT AT… NA       1       
#>  2 AB12345 AB12345…     2 EOR0102  EOR01… EORT… NA     QUITE … NA       3       
#>  3 AB12345 AB12345…     3 EOR0103  EOR01… EORT… NA     NOT AT… NA       1       
#>  4 AB12345 AB12345…     4 EOR0104  EOR01… EORT… NA     NA      NA       NA      
#>  5 AB12345 AB12345…     5 EOR0105  EOR01… EORT… NA     NOT AT… NA       1       
#>  6 AB12345 AB12345…     6 EOR0106  EOR01… EORT… NA     A LITT… NA       2       
#>  7 AB12345 AB12345…     7 EOR0107  EOR01… EORT… NA     VERY M… NA       4       
#>  8 AB12345 AB12345…     8 EOR0108  EOR01… EORT… NA     QUITE … NA       3       
#>  9 AB12345 AB12345…     9 EOR0109  EOR01… EORT… NA     NOT AT… NA       1       
#> 10 AB12345 AB12345…    10 EOR0110  EOR01… EORT… NA     NA      NA       NA      
#> # ℹ 1,761 more rows
#> # ℹ 7 more variables: QSSTRESU <chr>, QSSTAT <chr>, QSREASND <chr>,
#> #   VISITNUM <dbl>, VISIT <fct>, QSDTC <dttm>, QSEVLINT <chr>
df <- dplyr::left_join(
  adsl,
  qs,
  by = c("STUDYID", "USUBJID"),
  multiple = "all"
) |>
  dplyr::mutate(
    AVISIT = VISIT,
    PARAMCD = QSTESTCD,
    AVISITN = VISITNUM
  ) |>
  dplyr::mutate(ADTM = random.cdisc.data:::get_random_dates_between(TRTSDTM, TRTEDTM, AVISITN))
df
#> # A tibble: 1,771 × 74
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  2 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  3 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  4 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  5 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  6 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  7 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  8 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  9 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#> 10 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#> # ℹ 1,761 more rows
#> # ℹ 63 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …
adqlqc1 <- random.cdisc.data:::prep_adqlqc(df = qs)
adqlqc1
#> # A tibble: 1,771 × 22
#>    STUDYID USUBJID  QSSEQ QSTESTCD QSTEST QSCAT QSSCAT QSORRES QSORRESU QSSTRESC
#>    <chr>   <chr>    <int> <chr>    <chr>  <chr> <chr>  <chr>   <chr>    <chr>   
#>  1 AB12345 AB12345…     1 EOR0101  EOR01… EORT… NA     NOT AT… NA       1       
#>  2 AB12345 AB12345…     2 EOR0102  EOR01… EORT… NA     QUITE … NA       3       
#>  3 AB12345 AB12345…     3 EOR0103  EOR01… EORT… NA     NOT AT… NA       1       
#>  4 AB12345 AB12345…     4 EOR0104  EOR01… EORT… NA     NA      NA       NA      
#>  5 AB12345 AB12345…     5 EOR0105  EOR01… EORT… NA     NOT AT… NA       1       
#>  6 AB12345 AB12345…     6 EOR0106  EOR01… EORT… NA     A LITT… NA       2       
#>  7 AB12345 AB12345…     7 EOR0107  EOR01… EORT… NA     VERY M… NA       4       
#>  8 AB12345 AB12345…     8 EOR0108  EOR01… EORT… NA     QUITE … NA       3       
#>  9 AB12345 AB12345…     9 EOR0109  EOR01… EORT… NA     NOT AT… NA       1       
#> 10 AB12345 AB12345…    10 EOR0110  EOR01… EORT… NA     NA      NA       NA      
#> # ℹ 1,761 more rows
#> # ℹ 12 more variables: QSSTRESU <chr>, QSSTAT <chr>, QSREASND <chr>,
#> #   VISITNUM <dbl>, VISIT <fct>, QSDTC <dttm>, QSEVLINT <chr>, PARAMCD <chr>,
#> #   PARAM <chr>, PARCAT2 <chr>, PARCAT1N <chr>, PARCAT1 <chr>
df_scales <- random.cdisc.data:::calc_scales(df)
df_scales
#> # A tibble: 2,716 × 80
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  2 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  3 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  4 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  5 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  6 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  7 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  8 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  9 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> # ℹ 2,706 more rows
#> # ℹ 69 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …
adqlqc <- random.cdisc.data:::derv_chgcat1(dataset = adqlqc |> dplyr::select(-CHGCAT1))
adqlqc
#> # A tibble: 2,860 × 50
#>    STUDYID USUBJID   SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY TRT01P
#>    <chr>   <chr>     <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct> 
#>  1 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  2 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  3 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  4 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  5 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  6 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  7 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  8 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#>  9 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#> 10 AB12345 AB12345-… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     C: Co…
#> # ℹ 2,850 more rows
#> # ℹ 39 more variables: TRT01A <fct>, REGION1 <fct>, ITTFL <fct>, SAFFL <fct>,
#> #   TRTSDTM <dttm>, TRTEDTM <dttm>, QSSEQ <int>, QSCAT <fct>, QSSCAT <fct>,
#> #   QSDTC <dttm>, QSSTAT <fct>, QSSTRESC <fct>, QSSTRESU <fct>, QSORRES <fct>,
#> #   QSORRESU <fct>, QSTEST <fct>, QSTESTCD <fct>, QSREASND <fct>,
#> #   QSEVLINT <fct>, VISIT <fct>, VISITNUM <int>, ASEQ <int>, PARAM <fct>,
#> #   PARAMCD <fct>, PARCAT1 <fct>, PARCAT1N <dbl>, PARCAT2 <fct>, AVAL <dbl>, …
compliance_data <- random.cdisc.data:::comp_derv(adqlqc, 80, 2)
compliance_data
#> # A tibble: 144 × 12
#>    STUDYID USUBJID      PARCAT1 AVISIT AVISITN ADTM                  ADY PARAMCD
#>    <chr>   <chr>        <fct>   <fct>    <int> <dttm>              <int> <chr>  
#>  1 AB12345 AB12345-BRA… EORTC … CYCLE…       1 2021-01-18 22:43:53     2 CO028A…
#>  2 AB12345 AB12345-BRA… EORTC … CYCLE…       2 2021-02-15 22:43:53    30 CO028A…
#>  3 AB12345 AB12345-BRA… EORTC … CYCLE…       3 2021-03-15 22:43:53    58 CO028A…
#>  4 AB12345 AB12345-BRA… EORTC … CYCLE…       4 2021-04-12 22:43:53    86 CO028A…
#>  5 AB12345 AB12345-BRA… EORTC … CYCLE…       2 2020-02-19 10:19:08    30 CO028A…
#>  6 AB12345 AB12345-BRA… EORTC … CYCLE…       3 2020-03-18 10:19:08    58 CO028A…
#>  7 AB12345 AB12345-BRA… EORTC … CYCLE…       4 2020-04-15 10:19:08    86 CO028A…
#>  8 AB12345 AB12345-CHN… EORTC … SCREE…      -1 2020-04-25 02:28:45    -7 CO028A…
#>  9 AB12345 AB12345-CHN… EORTC … CYCLE…       0 2020-05-02 02:28:45     1 CO028A…
#> 10 AB12345 AB12345-CHN… EORTC … CYCLE…       1 2020-05-03 02:28:45     2 CO028A…
#> # ℹ 134 more rows
#> # ℹ 4 more variables: PARAM <chr>, PARCAT2 <chr>, AVAL <dbl>, AVALC <chr>