Generate Anthropometric Measurements for Males and Females.

Anthropometric measurements are randomly generated using normal approximation. The default mean and standard deviation values used are based on US National Health Statistics for adults aged 20 years or over. The measurements are generated in same units as provided to the function.

Usage

h_anthropometrics_by_sex(
  df,
  seed = 1,
  id_var = "USUBJID",
  sex_var = "SEX",
  sex_var_level_male = "M",
  male_weight_in_kg = list(mean = 90.6, sd = 44.9),
  female_weight_in_kg = list(mean = 77.5, sd = 46.2),
  male_height_in_m = list(mean = 1.75, sd = 0.14),
  female_height_in_m = list(mean = 1.61, sd = 0.24)
)

Arguments

df

(data.frame)
Analysis dataset.

seed

(numeric)
Seed to use for reproducible random number generation.

id_var

(character)
Patient identifier variable name.

sex_var

(character)
Name of variable representing sex of patient.

sex_var_level_male

(character)
Level of sex_var representing males.

male_weight_in_kg

(named list)
List of means and SDs of male weights in kilograms.

female_weight_in_kg

(named list)
List of means and SDs of female weights in kilograms.

male_height_in_m

(named list)
List of means and SDs of male heights in metres.

female_height_in_m

(named list)
list of means and SDs of female heights in metres.

Value

a dataframe with anthropometric measurements for each subject in analysis dataset.

Details

One record per subject.