NA: replace values

Replace column values with NAs

Usage

mutate_na(ds, na_vars = NULL, na_percentage = 0.05)

Arguments

ds

(data.frame) Any data set

na_vars

(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

na_percentage

(numeric) Default percentage of values to be replaced by NA