Protocol Deviations Analysis Dataset (ADDV)

Function for generating random Protocol Deviations Analysis Dataset for a given Subject-Level Analysis Dataset.

Usage

raddv(
  adsl,
  max_n_dv = 3L,
  p_dv = 0.15,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(ASTDT = c(seed = 1234, percentage = 0.1), DVCAT = c(seed = 1234,
    percentage = 0.1)),
  cached = FALSE
)

Arguments

adsl

(data.frame)
Subject-Level Analysis Dataset (ADSL).

max_n_dv

(integer)
Maximum number of deviations per patient. Defaults to 3.

p_dv

(proportion)
Probability of a patient having protocol deviations.

lookup

(data.frame)
Additional parameters.

seed

(numeric)
Seed to use for reproducible random number generation.

na_percentage

(proportion)
Default percentage of values to be replaced by NA.

na_vars

(list)
A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements:

seed (numeric)
The seed to be used for this element - can be NA.

percentage (proportion)
Percentage of elements to be replaced with NA. If NA, na_percentage is used as a default.

cached

boolean whether the cached ADDV data caddv should be returned or new data should be generated. If set to TRUE then the other arguments to raddv will be ignored.

Value

data.frame

Details

One record per each record in the corresponding SDTM domain.

Keys: STUDYID, USUBJID, ASTDT, DVTERM, DVSEQ

Examples

adsl <- radsl(N = 10, seed = 1, study_duration = 2)

addv <- raddv(adsl, seed = 2)
addv
#> # A tibble: 9 × 66
#>   STUDYID USUBJID     SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>   <chr>   <chr>       <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#> 1 AB12345 AB12345-BR… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> 2 AB12345 AB12345-BR… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> 3 AB12345 AB12345-BR… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> 4 AB12345 AB12345-CH… id-5   CHN-3     36 YEARS F     ASIAN NOT H… CHN     N    
#> 5 AB12345 AB12345-CH… id-5   CHN-3     36 YEARS F     ASIAN NOT H… CHN     N    
#> 6 AB12345 AB12345-CH… id-5   CHN-3     36 YEARS F     ASIAN NOT H… CHN     N    
#> 7 AB12345 AB12345-US… id-3   USA-13    35 YEARS F     AMER… NOT H… USA     N    
#> 8 AB12345 AB12345-US… id-3   USA-13    35 YEARS F     AMER… NOT H… USA     N    
#> 9 AB12345 AB12345-US… id-3   USA-13    35 YEARS F     AMER… NOT H… USA     N    
#> # ℹ 55 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>,
#> #   AP01SDTM <dttm>, AP01EDTM <dttm>, AP02SDTM <dttm>, AP02EDTM <dttm>, …