ECG Analysis Dataset (ADEG)

Function for generating random dataset from ECG Analysis Dataset for a given Subject-Level Analysis Dataset.

Usage

radeg(
  adsl,
  egcat = c("INTERVAL", "INTERVAL", "MEASUREMENT", "FINDING"),
  param = c("QT Duration", "RR Duration", "Heart Rate", "ECG Interpretation"),
  paramcd = c("QT", "RR", "HR", "ECGINTP"),
  paramu = c("msec", "msec", "beats/min", ""),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_eg = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(ABLFL = c(1235, 0.1), BASE = c(NA, 0.1), BASEC = c(NA, 0.1), CHG =
    c(1234, 0.1), PCHG = c(1234, 0.1)),
  cached = FALSE
)

Arguments

adsl

(data.frame)
Subject-Level Analysis Dataset (ADSL).

egcat

(character vector)
EG category values.

param

(character vector)
Parameter values.

paramcd

(character vector)
Parameter code values.

paramu

(character vector)
Parameter unit values.

visit_format

(character)
Type of visit. Options are "WEEK" and "CYCLE".

n_assessments

(integer)
Number of weeks or cycles.

n_days

(integer)
Number of days in each cycle (only used if visit_format is "CYCLE").

max_n_eg

(integer)
Maximum number of EG results per patient. Defaults to 10.

lookup

(data.frame)
Additional parameters.

seed

(numeric)
Seed to use for reproducible random number generation.

na_percentage

(proportion)
Default percentage of values to be replaced by NA.

na_vars

(list)
A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements:

seed (numeric)
The seed to be used for this element - can be NA.

percentage (proportion)
Percentage of elements to be replaced with NA. If NA, na_percentage is used as a default.

cached

boolean whether the cached ADEG data cadeg should be returned or new data should be generated. If set to TRUE then the other arguments to radeg will be ignored.

Value

data.frame

Details

One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, BASETYPE, AVISITN, ATPTN, DTYPE, ADTM, EGSEQ, ASPID

Author

tomlinsj, npaszty, Xuefeng Hou, dipietrc

Examples

adsl <- radsl(N = 10, seed = 1, study_duration = 2)

adeg <- radeg(adsl, visit_format = "WEEK", n_assessments = 7L, seed = 2)
adeg
#> # A tibble: 420 × 88
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  2 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  3 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  4 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  5 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  6 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  7 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  8 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  9 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> # ℹ 410 more rows
#> # ℹ 77 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …

adeg <- radeg(adsl, visit_format = "CYCLE", n_assessments = 2L, seed = 2)
adeg
#> # A tibble: 500 × 88
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  2 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  3 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  4 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  5 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  6 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  7 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  8 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#>  9 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> # ℹ 490 more rows
#> # ℹ 77 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>, …