Laboratory Data Analysis Dataset (ADLB)

Function for generating random dataset from Laboratory Data Analysis Dataset for a given Subject-Level Analysis Dataset.

radlb(
  ADSL,
  lbcat = c("CHEMISTRY", "CHEMISTRY", "IMMUNOLOGY"),
  param = c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement",
    "Immunoglobulin A Measurement"),
  paramcd = c("ALT", "CRP", "IGA"),
  paramu = c("U/L", "mg/L", "g/L"),
  aval_mean = c(20, 1, 2),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_lbs = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(LOQFL = c(NA, 0.1), ABLFL2 = c(1234, 0.1), ABLFL = c(1235, 0.1), BASE2 =
    c(NA, 0.1), BASE = c(NA, 0.1), CHG2 = c(1235, 0.1), PCHG2 = c(1235, 0.1), CHG =
    c(1234, 0.1), PCHG = c(1234, 0.1)),
  cached = FALSE
)

Arguments

ADSL

Subject-Level Analysis Dataset (ADSL).

lbcat

As character vector of lb category values.

param

As character string. list of parameter values.

paramcd

As character string. list of parameter code values.

paramu

As character vector of parameter unit values.

aval_mean

As numerical vector of appropriate mean values for each lab test.

visit_format

Type of visit either "WEEK" or "CYCLE".

n_assessments

Number of weeks or cycles.

n_days

Number of days within cycle.

max_n_lbs

As numeric. maximum number of labs.

lookup

control lookup process.

seed

Seed for random number generation.

na_percentage

(numeric) Default percentage of values to be replaced by NA

na_vars

(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached

boolean whether the cached ADLB data cadlb should be returned or new data should be generated. If set to TRUE then the other arguments to radlb will be ignored.

Value

data.frame

Details

One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, BASETYPE, AVISITN, ATPTN, DTYPE, ADTM, LBSEQ, ASPID.

Author

tomlinsj, npaszty, Xuefeng Hou

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radlb(ADSL, visit_format = "WEEK", n_assessments = 7L, seed = 2)
#> # A tibble: 270 × 87
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  2 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  3 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  4 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  5 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  6 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  7 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  8 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  9 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#> # … with 260 more rows, and 76 more variables: INVID <chr>, INVNAM <chr>,
#> #   ARM <fct>, ARMCD <fct>, ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>,
#> #   TRT01A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   EOSSTT <fct>, EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, …
radlb(ADSL, visit_format = "CYCLE", n_assessments = 2L, seed = 2)
#> # A tibble: 330 × 87
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  2 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  3 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  4 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  5 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  6 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  7 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  8 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#>  9 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… HISPA… BRA     N    
#> # … with 320 more rows, and 76 more variables: INVID <chr>, INVNAM <chr>,
#> #   ARM <fct>, ARMCD <fct>, ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>,
#> #   TRT01A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   EOSSTT <fct>, EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, …