Pharmacokinetics Analysis Dataset

Pharmacokinetics Analysis Dataset

Usage

radpc(
  ADSL,
  avalu = "ug/mL",
  constants = c(D = 100, ka = 0.8, ke = 1),
  duration = 2,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)

Arguments

ADSL

Subject-Level Analysis Dataset (ADSL).

avalu

(`string`)
Analysis value unit value

constants

(`character vector`)
Constant parameters to be used in the PK equation for creating analysis values.

duration

(`numeric`)
Duration in number of days.

seed

Seed for random number generation.

na_percentage

(numeric) Default percentage of values to be replaced by NA

na_vars

(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached

boolean whether the cached ADSL data cadsl should be returned or new data should be generated. If set to TRUE then the other arguments to radsl will be ignored.

Details

One record per per study per subject per parameter per time point

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPC <- radpc(ADSL, seed = 2)
ADPC <- radpc(ADSL, seed = 2, duration = 3)