PK parameters dataset

PK parameters dataset

radpp(
  ADSL,
  ppcat = c("Plasma Drug X", "Plasma Drug Y", "Metabolite Drug X", "Metabolite Drug Y"),
  ppspec = c("Plasma", "Plasma", "Plasma", "Matrix of PD", "Matrix of PD", "Urine",
    "Urine", "Urine", "Urine"),
  paramcd = c("AUCIFO", "CMAX", "CLO", "RMAX", "TON", "RENALCL", "RENALCLD", "RCAMINT",
    "RCPCINT"),
  param = c("AUC Infinity Obs", "Max Conc", "Total CL Obs", "Time of Maximum Response",
    "Time to Onset", "Renal CL", "Renal CL Norm by Dose", "Amt Rec from T1 to T2",
    "Pct Rec from T1 to T2"),
  paramu = c("day*ug/mL", "ug/mL", "ml/day/kg", "hr", "hr", "L/hr", "L/hr/mg", "mg",
    "%"),
  aval_mean = c(200, 30, 5, 10, 3, 0.05, 0.005, 1.5613, 15.65),
  visit_format = "CYCLE",
  n_days = 2L,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(AVAL = c(NA, 0.1)),
  cached = FALSE
)

Arguments

ADSL

Subject-Level Analysis Dataset (ADSL).

ppcat

(`character`)
category of parameter.

ppspec

(`string`)
specimen material type.

paramcd

(`character`)
parameter code.

param

(`character`)
parameters.

paramu

(`character`)
Analysis value unit value.

aval_mean

(`vector`)
the mean for corresponding parameter.

visit_format

(`string`)
the visit format to produce AVISIT in the dataset.

n_days

(`integer`)
the number days that will be included to produce AVISIT in the dataset.

seed

Seed for random number generation.

na_percentage

(numeric) Default percentage of values to be replaced by NA

na_vars

(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached

boolean whether the cached ADSL data cadsl should be returned or new data should be generated. If set to TRUE then the other arguments to radsl will be ignored.

Details

One record per per study per subject per parameter category per parameter per visit.

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
ADPP <- radpp(ADSL, seed = 2)