Generate a random ADSL dataset

Function for generating random Subject-Level Analysis Dataset

radsl(N = 400, seed = NULL)

Arguments

N
number of patients

seed
seed for random number generation

Value

a data frame containing following attributes is returned: [,1] SUBJID (Subject Identification Number),
[,2] STUDYID (Study Identification Number),
[,3] SITEID (Site Identification Number),
[,4] USUBJID (Unique Subject Identifier),
[,5] AGE,
[,6] SEX,
[,7] ARMCD (Planned Arme Code),
[,8] COUNTRY,
[,9] RACE,
[,10] STRATA1 (Stratification Factor 1),
[,11] STRATA2 (Stratification Factor 2),
[,12] BMRKR1 (Cont. Biomarker 1),
[,13] BMRKR2 (Cat. Biomarker 2),
[,14] ARM (Description of Planned Arm),
[,15] ACTARM (Description of Actual Arm).

Details

The Subject-Level Analysis Data set (ADSL) is used to provide the variables that describe attributes of a subject. ADSL is a source for subject-level variables used in other analysis data sets, such as population flags and treatment variables. There is only one ADSL per study. ADSL and its related metadata are required in a CDISC-based submission of data from a clinical trial even if no other analysis data sets are submitted.

Examples

ADSL <- radsl()
head(ADSL)

#> # A tibble: 6 x 15
#>   SUBJID STUDYID SITEID USUBJID   AGE SEX   ARMCD COUNTRY RACE  STRATA1 STRATA2
#>   <chr>  <chr>   <fct>  <chr>   <dbl> <fct> <fct> <fct>   <fct> <fct>   <fct>  
#> 1 id-1   AB12345 XYZ2   AB1234~    51 M     ARM C CUW     OTHER C       S1     
#> 2 id-2   AB12345 XYZ1   AB1234~    20 M     ARM A CUW     WHITE A       S1     
#> 3 id-3   AB12345 XYZ2   AB1234~    36 F     ARM C BES     AMER~ B       S2     
#> 4 id-4   AB12345 XYZ2   AB1234~    62 M     ARM B AFG     UNKN~ A       S2     
#> 5 id-5   AB12345 XYZ1   AB1234~    34 M     ARM A SXM     WHITE A       S1     
#> 6 id-6   AB12345 XYZ2   AB1234~    22 M     ARM C BES     MULT~ B       S1     
#> # ... with 4 more variables: BMRKR1 <dbl>, BMRKR2 <fct>, ARM <fct>,
#> #   ACTARM <fct>