Subcategory Analysis Dataset (ADSUB)

Function for generating random dataset from Subcategory Analysis Dataset for a given Subject-Level Analysis Dataset

radsub(
  ADSL,
  param = c("Baseline Weight", "Baseline Height", "Baseline BMI", "Baseline ECOG",
    "Baseline Biomarker Mutation"),
  paramcd = c("BWGHTSI", "BHGHTSI", "BBMISI", "BECOG", "BBMRKR1"),
  seed = NULL,
  na_percentage = 0,
  na_vars = list(),
  cached = FALSE
)

Arguments

ADSL

Subject-Level Analysis Dataset (ADSL).

param

param character vector with parameter value strings.

paramcd

paramcd character vector with parameter code values.

seed

Seed for random number generation.

na_percentage

(numeric) Default percentage of values to be replaced by NA

na_vars

(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached

boolean whether the cached ADSUB data cadsub should be returned or new data should be generated. If set to TRUE then the other arguments to radsub will be ignored.

Value

data.frame

Details

One record per subject

Keys: STUDYID, USUBJID, PARAMCD, AVISITN, ADTM, SRCSEQ

Author

tomlinsj, npaszty, Xuefeng Hou, dipietrc

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radsub(ADSL, seed = 2)
#> # A tibble: 50 × 49
#>    STUDYID USUBJID  SUBJID SITEID   AGE SEX   RACE    COUNTRY DTHFL INVID INVNAM
#>    <chr>   <chr>    <chr>  <chr>  <int> <fct> <fct>   <fct>   <fct> <chr> <chr> 
#>  1 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  2 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  3 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  4 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  5 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  6 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#>  7 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#>  8 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#>  9 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#> 10 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#> # … with 40 more rows, and 38 more variables: ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>,
#> #   BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>,
#> #   EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>,
#> #   DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>,
#> #   LSTALVDT <date>, study_duration_secs <dbl>, PARAM <fct>, PARAMCD <fct>, …