Tumor Response Analysis Dataset (ADTR)

Function for generating random dataset from Tumor Response Analysis Dataset for a given Subject-Level Analysis Dataset.

radtr(
  ADSL,
  param = c("Sum of Longest Diameter by Investigator"),
  paramcd = c("SLDINV"),
  seed = NULL,
  cached = FALSE,
  ...
)

Arguments

ADSL

Subject-Level Analysis Dataset (ADSL).

param

As character string. list of parameter values.

paramcd

As character string. list of parameter code values.

seed

Seed for random number generation.

cached

boolean whether the cached ADTR data cadtr should be returned or new data should be generated. If set to TRUE then the other arguments to radtr will be ignored.

...

parameters from radrs

Value

data.frame

Details

One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, BASETYPE, AVISITN, DTYPE

Author

tomlinsj, npaszty, Xuefeng Hou, dipietrc

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radtr(ADSL, seed = 2)
#>    STUDYID             USUBJID SUBJID SITEID AGE SEX
#> 1  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 2  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 3  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 4  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 5  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 6  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 7  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 8  AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 9  AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 10 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 11 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 12 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 13 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 14 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 15 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 16 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 17 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 18 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 19 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 20 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 21 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 22 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 23 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 24 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 25 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 26 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 27 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 28 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 29 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 30 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 31 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 32 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 33 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 34 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 35 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 36 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 37 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 38 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 39 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 40 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 41 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 42 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 43 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 44 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 45 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 46 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 47 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 48 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 49 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 50 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 51 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 52 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 53 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 54 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 55 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 56 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 57 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 58 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 59 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 60 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 61 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 62 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 63 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 64 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 65 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 66 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 67 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 68 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 69 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 70 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#>                                RACE                 ETHNIC COUNTRY DTHFL
#> 1         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 2         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 3         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 4         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 5         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 6         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 7         BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     BRA     N
#> 8                             ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 9                             ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 10                            ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 11                            ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 12                            ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 13                            ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 14                            ASIAN NOT HISPANIC OR LATINO     BRA     N
#> 15                            WHITE                UNKNOWN     CHN     N
#> 16                            WHITE                UNKNOWN     CHN     N
#> 17                            WHITE                UNKNOWN     CHN     N
#> 18                            WHITE                UNKNOWN     CHN     N
#> 19                            WHITE                UNKNOWN     CHN     N
#> 20                            WHITE                UNKNOWN     CHN     N
#> 21                            WHITE                UNKNOWN     CHN     N
#> 22        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 23        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 24        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 25        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 26        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 27        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 28        BLACK OR AFRICAN AMERICAN NOT HISPANIC OR LATINO     CHN     Y
#> 29                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 30                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 31                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 32                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 33                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 34                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 35                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 36                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 37                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 38                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 39                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 40                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 41                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 42                            ASIAN NOT HISPANIC OR LATINO     CHN     N
#> 43                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 44                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 45                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 46                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 47                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 48                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 49                            ASIAN NOT HISPANIC OR LATINO     JPN     N
#> 50        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 51        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 52        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 53        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 54        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 55        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 56        BLACK OR AFRICAN AMERICAN     HISPANIC OR LATINO     RUS     N
#> 57                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 58                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 59                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 60                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 61                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 62                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 63                            ASIAN NOT HISPANIC OR LATINO     RUS     Y
#> 64 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#> 65 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#> 66 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#> 67 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#> 68 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#> 69 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#> 70 AMERICAN INDIAN OR ALASKA NATIVE NOT HISPANIC OR LATINO     USA     N
#>            INVID         INVNAM            ARM ARMCD         ACTARM ACTARMCD
#> 1   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 2   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 3   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 4   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 5   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 6   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 7   INV ID BRA-1  Dr. BRA-1 Doe C: Combination ARM C C: Combination    ARM C
#> 8   INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 9   INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 10  INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 11  INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 12  INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 13  INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 14  INV ID BRA-9  Dr. BRA-9 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 15 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 16 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 17 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 18 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 19 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 20 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 21 INV ID CHN-11 Dr. CHN-11 Doe C: Combination ARM C C: Combination    ARM C
#> 22 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 23 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 24 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 25 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 26 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 27 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 28 INV ID CHN-11 Dr. CHN-11 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 29  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 30  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 31  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 32  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 33  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 34  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 35  INV ID CHN-3  Dr. CHN-3 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 36  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 37  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 38  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 39  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 40  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 41  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 42  INV ID CHN-3  Dr. CHN-3 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 43  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 44  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 45  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 46  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 47  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 48  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 49  INV ID JPN-4  Dr. JPN-4 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 50  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 51  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 52  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 53  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 54  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 55  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 56  INV ID RUS-1  Dr. RUS-1 Doe     B: Placebo ARM B     B: Placebo    ARM B
#> 57  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 58  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 59  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 60  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 61  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 62  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 63  INV ID RUS-1  Dr. RUS-1 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 64 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 65 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 66 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 67 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 68 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 69 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#> 70 INV ID USA-13 Dr. USA-13 Doe      A: Drug X ARM A      A: Drug X    ARM A
#>          REGION1 STRATA1 STRATA2    BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL
#> 1  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 2  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 3  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 4  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 5  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 6  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 7  South America       A      S1  6.922043   HIGH     Y     Y        Y       N
#> 8  South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 9  South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 10 South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 11 South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 12 South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 13 South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 14 South America       B      S2 11.277028 MEDIUM     Y     Y        Y       Y
#> 15          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 16          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 17          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 18          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 19          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 20          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 21          Asia       B      S1  8.926907    LOW     Y     Y        N       N
#> 22          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 23          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 24          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 25          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 26          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 27          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 28          Asia       C      S2  4.084350   HIGH     Y     Y        Y       Y
#> 29          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 30          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 31          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 32          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 33          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 34          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 35          Asia       A      S1  1.777905    LOW     Y     Y        N       Y
#> 36          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 37          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 38          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 39          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 40          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 41          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 42          Asia       A      S2  2.619796    LOW     Y     Y        Y       N
#> 43          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 44          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 45          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 46          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 47          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 48          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 49          Asia       B      S1  7.459420 MEDIUM     Y     Y        Y       N
#> 50       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 51       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 52       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 53       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 54       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 55       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 56       Eurasia       A      S1  9.952892 MEDIUM     Y     Y        Y       N
#> 57       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 58       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 59       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 60       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 61       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 62       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 63       Eurasia       B      S1  9.342782 MEDIUM     Y     Y        N       N
#> 64 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#> 65 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#> 66 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#> 67 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#> 68 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#> 69 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#> 70 North America       C      S1  7.254058    LOW     Y     Y        Y       N
#>        RANDDT             TRTSDTM             TRTEDTM       EOSSTT       EOTSTT
#> 1  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 2  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 3  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 4  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 5  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 6  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 7  2021-01-17 2021-01-17 22:43:53 2023-01-18 10:22:17    COMPLETED    COMPLETED
#> 8  2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 9  2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 10 2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 11 2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 12 2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 13 2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 14 2020-01-18 2020-01-21 10:19:08                <NA>      ONGOING      ONGOING
#> 15 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 16 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 17 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 18 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 19 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 20 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 21 2020-04-29 2020-05-02 02:28:45 2022-05-02 14:07:09    COMPLETED    COMPLETED
#> 22 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 23 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 24 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 25 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 26 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 27 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 28 2020-02-23 2020-02-26 18:15:59 2021-01-22 14:18:49 DISCONTINUED DISCONTINUED
#> 29 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 30 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 31 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 32 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 33 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 34 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 35 2019-03-24 2019-03-28 12:07:16 2021-03-27 23:45:40    COMPLETED    COMPLETED
#> 36 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 37 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 38 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 39 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 40 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 41 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 42 2020-06-24 2020-06-27 04:48:00 2020-11-06 15:43:55 DISCONTINUED DISCONTINUED
#> 43 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 44 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 45 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 46 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 47 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 48 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 49 2019-09-22 2019-09-22 02:25:18 2021-09-21 14:03:42    COMPLETED    COMPLETED
#> 50 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 51 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 52 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 53 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 54 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 55 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 56 2020-08-27 2020-08-28 20:42:09 2022-08-29 08:20:33    COMPLETED    COMPLETED
#> 57 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 58 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 59 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 60 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 61 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 62 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 63 2020-10-24 2020-10-26 22:16:26 2021-01-30 03:55:51 DISCONTINUED DISCONTINUED
#> 64 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#> 65 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#> 66 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#> 67 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#> 68 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#> 69 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#> 70 2019-10-16 2019-10-17 17:56:02 2021-10-17 05:34:26    COMPLETED    COMPLETED
#>         EOSDT EOSDY       DCSREAS      DTHDT       DTHCAUS        DTHCAT
#> 1  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 2  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 3  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 4  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 5  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 6  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 7  2023-01-18   731          <NA>       <NA>          <NA>          <NA>
#> 8        <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 9        <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 10       <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 11       <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 12       <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 13       <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 14       <NA>    NA          <NA>       <NA>          <NA>          <NA>
#> 15 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 16 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 17 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 18 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 19 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 20 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 21 2022-05-02   731          <NA>       <NA>          <NA>          <NA>
#> 22 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 23 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 24 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 25 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 26 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 27 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 28 2021-01-22   331         DEATH 2021-01-23 ADVERSE EVENT ADVERSE EVENT
#> 29 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 30 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 31 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 32 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 33 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 34 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 35 2021-03-27   731          <NA>       <NA>          <NA>          <NA>
#> 36 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 37 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 38 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 39 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 40 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 41 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 42 2020-11-06   133 ADVERSE EVENT       <NA>          <NA>          <NA>
#> 43 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 44 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 45 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 46 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 47 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 48 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 49 2021-09-21   731          <NA>       <NA>          <NA>          <NA>
#> 50 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 51 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 52 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 53 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 54 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 55 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 56 2022-08-29   731          <NA>       <NA>          <NA>          <NA>
#> 57 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 58 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 59 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 60 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 61 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 62 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 63 2021-01-30    96         DEATH 2021-02-24 ADVERSE EVENT ADVERSE EVENT
#> 64 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#> 65 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#> 66 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#> 67 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#> 68 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#> 69 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#> 70 2021-10-17   731          <NA>       <NA>          <NA>          <NA>
#>    LDDTHELD LDDTHGR1   LSTALVDT DTHADY study_duration_secs AGEGR1   TRTDURD
#> 1        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 2        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 3        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 4        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 5        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 6        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 7        NA     <NA> 2023-02-15     NA            63113904    <65 730.48499
#> 8        NA     <NA>       <NA>     NA            63113904    <65        NA
#> 9        NA     <NA>       <NA>     NA            63113904    <65        NA
#> 10       NA     <NA>       <NA>     NA            63113904    <65        NA
#> 11       NA     <NA>       <NA>     NA            63113904    <65        NA
#> 12       NA     <NA>       <NA>     NA            63113904    <65        NA
#> 13       NA     <NA>       <NA>     NA            63113904    <65        NA
#> 14       NA     <NA>       <NA>     NA            63113904    <65        NA
#> 15       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 16       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 17       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 18       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 19       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 20       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 21       NA     <NA> 2022-05-25     NA            63113904    <65 730.48499
#> 22        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 23        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 24        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 25        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 26        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 27        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 28        1     <=30 2021-01-23    332            63113904    <65 330.83529
#> 29       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 30       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 31       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 32       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 33       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 34       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 35       NA     <NA> 2021-04-10     NA            63113904    <65 730.48499
#> 36       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 37       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 38       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 39       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 40       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 41       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 42       NA     <NA> 2020-11-17     NA            63113904    <65 132.45549
#> 43       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 44       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 45       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 46       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 47       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 48       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 49       NA     <NA> 2021-10-03     NA            63113904    <65 730.48499
#> 50       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 51       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 52       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 53       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 54       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 55       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 56       NA     <NA> 2022-09-10     NA            63113904    <65 730.48499
#> 57       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 58       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 59       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 60       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 61       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 62       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 63       25     <=30 2021-02-24    121            63113904    <65  95.23569
#> 64       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#> 65       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#> 66       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#> 67       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#> 68       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#> 69       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#> 70       NA     <NA> 2021-11-15     NA            63113904    <65 730.48499
#>    DCSREAS_GRP AEWITHFL                                   PARAM PARAMCD
#> 1         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 2         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 3         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 4         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 5         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 6         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 7         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 8         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 9         <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 10        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 11        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 12        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 13        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 14        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 15        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 16        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 17        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 18        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 19        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 20        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 21        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 22  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 23  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 24  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 25  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 26  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 27  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 28  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 29        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 30        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 31        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 32        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 33        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 34        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 35        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 36      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 37      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 38      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 39      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 40      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 41      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 42      Safety        Y Sum of Longest Diameter by Investigator  SLDINV
#> 43        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 44        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 45        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 46        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 47        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 48        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 49        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 50        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 51        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 52        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 53        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 54        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 55        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 56        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 57  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 58  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 59  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 60  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 61  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 62  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 63  Non-Safety        N Sum of Longest Diameter by Investigator  SLDINV
#> 64        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 65        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 66        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 67        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 68        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 69        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#> 70        <NA>     <NA> Sum of Longest Diameter by Investigator  SLDINV
#>                   AVISIT AVISITN                ADTM  ADY   DTYPE      AVAL
#> 1       END OF INDUCTION     999 2021-12-31 00:00:00  348          91.02365
#> 2  POST-BASELINE MINIMUM     999 2021-12-31 00:00:00  348 MINIMUM  91.02365
#> 3              SCREENING      -1 2020-10-09 00:00:00 -100         124.85139
#> 4               BASELINE       0 2021-01-17 22:43:53    0                NA
#> 5          CYCLE 2 DAY 1       2 2021-08-08 00:00:00  203         133.13259
#> 6          CYCLE 4 DAY 1       4 2021-10-07 00:00:00  263         188.27147
#> 7              FOLLOW UP     999 2021-10-16 00:00:00  272         118.57282
#> 8       END OF INDUCTION     999 2021-04-28 00:00:00  463          96.35273
#> 9          CYCLE 4 DAY 1       4                <NA>   NA         200.14856
#> 10             FOLLOW UP     999 2021-06-05 00:00:00  501         210.93728
#> 11 POST-BASELINE MINIMUM     999 2021-04-28 00:00:00  463 MINIMUM  96.35273
#> 12              BASELINE       0 2020-01-21 10:19:08    0                NA
#> 13         CYCLE 2 DAY 1       2 2021-05-14 00:00:00  479         184.17689
#> 14             SCREENING      -1 2019-10-13 00:00:00 -100         140.31087
#> 15         CYCLE 4 DAY 1       4 2022-03-16 00:00:00  683         125.40015
#> 16 POST-BASELINE MINIMUM     999 2021-02-14 00:00:00  288 MINIMUM  90.03459
#> 17             SCREENING      -1 2020-01-23 00:00:00 -100         128.90567
#> 18         CYCLE 2 DAY 1       2 2022-01-15 00:00:00  623         165.18704
#> 19             FOLLOW UP     999 2021-02-14 00:00:00  288          90.03459
#> 20              BASELINE       0 2020-05-02 02:28:45    0                NA
#> 21      END OF INDUCTION     999 2021-11-18 00:00:00  565         135.62122
#> 22              BASELINE       0 2020-02-26 18:15:59    0                NA
#> 23             SCREENING      -1 2019-11-18 00:00:00 -100         152.52540
#> 24             FOLLOW UP     999 2020-07-18 00:00:00  143         163.04543
#> 25         CYCLE 2 DAY 1       2 2020-10-20 00:00:00  237         122.36173
#> 26      END OF INDUCTION     999 2020-04-10 00:00:00   44         145.75018
#> 27 POST-BASELINE MINIMUM       2 2020-10-20 00:00:00  237 MINIMUM 122.36173
#> 28         CYCLE 4 DAY 1       4 2020-12-19 00:00:00  297         159.91349
#> 29         CYCLE 2 DAY 1       2 2020-03-11 00:00:00  349         189.10537
#> 30         CYCLE 4 DAY 1       4 2020-05-10 00:00:00  409         173.15369
#> 31              BASELINE       0 2019-03-28 12:07:16    0                NA
#> 32             FOLLOW UP     999 2020-05-29 00:00:00  428         107.69885
#> 33 POST-BASELINE MINIMUM     999 2020-05-29 00:00:00  428 MINIMUM 107.69885
#> 34      END OF INDUCTION     999 2020-04-20 00:00:00  389         181.57577
#> 35             SCREENING      -1 2018-12-18 00:00:00 -100         148.38832
#> 36         CYCLE 2 DAY 1       2 2020-08-06 00:00:00   40         133.99884
#> 37             FOLLOW UP     999 2020-11-02 00:00:00  128         204.66368
#> 38      END OF INDUCTION     999 2020-08-13 00:00:00   47          83.76241
#> 39              BASELINE       0 2020-06-27 04:48:00    0                NA
#> 40 POST-BASELINE MINIMUM     999 2020-08-13 00:00:00   47 MINIMUM  83.76241
#> 41         CYCLE 4 DAY 1       4 2020-10-05 00:00:00  100         108.83192
#> 42             SCREENING      -1 2020-03-19 00:00:00 -100         179.87954
#> 43 POST-BASELINE MINIMUM       2 2021-07-13 00:00:00  660 MINIMUM 138.39151
#> 44             FOLLOW UP     999 2021-03-23 00:00:00  548         200.43465
#> 45              BASELINE       0 2019-09-22 02:25:18    0                NA
#> 46      END OF INDUCTION     999 2020-05-04 00:00:00  225         198.01173
#> 47             SCREENING      -1 2019-06-14 00:00:00 -100         130.39820
#> 48         CYCLE 2 DAY 1       2 2021-07-13 00:00:00  660         138.39151
#> 49         CYCLE 4 DAY 1       4 2021-09-11 00:00:00  720         161.60085
#> 50             SCREENING      -1 2020-05-20 00:00:00 -100         114.49181
#> 51         CYCLE 2 DAY 1       2 2021-11-22 00:00:00  451         104.61988
#> 52         CYCLE 4 DAY 1       4 2022-01-21 00:00:00  511         112.40685
#> 53              BASELINE       0 2020-08-28 20:42:09    0                NA
#> 54 POST-BASELINE MINIMUM       2 2021-11-22 00:00:00  451 MINIMUM 104.61988
#> 55      END OF INDUCTION     999 2022-06-26 00:00:00  667         150.22938
#> 56             FOLLOW UP     999 2022-03-06 00:00:00  555         208.78071
#> 57              BASELINE       0 2020-10-26 22:16:26    0                NA
#> 58             SCREENING      -1 2020-07-18 00:00:00 -100         124.72154
#> 59      END OF INDUCTION     999 2020-12-15 00:00:00   50         140.57184
#> 60             FOLLOW UP     999 2021-01-29 00:00:00   95         127.51110
#> 61 POST-BASELINE MINIMUM     999 2021-01-29 00:00:00   95 MINIMUM 127.51110
#> 62         CYCLE 2 DAY 1       2 2020-11-09 00:00:00   14         182.23378
#> 63         CYCLE 4 DAY 1       4 2021-01-08 00:00:00   74         157.81794
#> 64 POST-BASELINE MINIMUM     999 2020-09-13 00:00:00  332 MINIMUM 137.35879
#> 65      END OF INDUCTION     999 2020-11-08 00:00:00  388         139.47497
#> 66         CYCLE 4 DAY 1       4 2020-05-29 00:00:00  225         210.26061
#> 67             FOLLOW UP     999 2020-09-13 00:00:00  332         137.35879
#> 68             SCREENING      -1 2019-07-09 00:00:00 -100         124.13405
#> 69         CYCLE 2 DAY 1       2 2020-03-30 00:00:00  165         178.19760
#> 70              BASELINE       0 2019-10-17 17:56:02    0                NA
#>               AVALC AVALU      BASE ABLFL          CHG        PCHG ONTRTFL
#> 1  91.0236527565802    mm 211.98904       -120.9653879 -57.0620951       Y
#> 2  91.0236527565802    mm 211.98904       -120.9653879 -57.0620951       Y
#> 3  124.851385571939    mm 211.98904                 NA          NA        
#> 4              <NA>    mm 211.98904     Y           NA          NA       Y
#> 5  133.132588399725    mm 211.98904        -78.8564523 -37.1983627       Y
#> 6  188.271465355528    mm 211.98904        -23.7175753 -11.1881139       Y
#> 7  118.572821190151    mm 211.98904        -93.4162195 -44.0665325       Y
#> 8  96.3527337970544    mm 178.07588        -81.7231420 -45.8923151       Y
#> 9   200.14856301978    mm 178.07588         22.0726872  12.3951024       Y
#> 10 210.937275562657    mm 178.07588         32.8613998  18.4535944       Y
#> 11 96.3527337970544    mm 178.07588        -81.7231420 -45.8923151       Y
#> 12             <NA>    mm 178.07588     Y           NA          NA       Y
#> 13 184.176894082022    mm 178.07588          6.1010183   3.4260779       Y
#> 14 140.310867187201    mm 178.07588                 NA          NA        
#> 15 125.400146831514    mm 154.74494        -29.3447960 -18.9633312       Y
#> 16 90.0345901510763    mm 154.74494        -64.7103527 -41.8174265       Y
#> 17 128.905670010183    mm 154.74494                 NA          NA        
#> 18 165.187043924422    mm 154.74494         10.4421010   6.7479433       Y
#> 19 90.0345901510763    mm 154.74494        -64.7103527 -41.8174265       Y
#> 20             <NA>    mm 154.74494     Y           NA          NA       Y
#> 21 135.621222256451    mm 154.74494        -19.1237206 -12.3582201       Y
#> 22             <NA>    mm 123.13540     Y           NA          NA       Y
#> 23 152.525397129667    mm 123.13540                 NA          NA        
#> 24 163.045432845081    mm 123.13540         39.9100312  32.4115004       Y
#> 25 122.361730034768    mm 123.13540         -0.7736716  -0.6283097       Y
#> 26 145.750175731918    mm 123.13540         22.6147741  18.3657776       Y
#> 27 122.361730034768    mm 123.13540         -0.7736716  -0.6283097       Y
#> 28 159.913485103959    mm 123.13540         36.7780834  29.8680013       Y
#> 29 189.105366972668    mm 122.78669         66.3186782  54.0112930       Y
#> 30 173.153693266196    mm 122.78669         50.3670045  41.0199225       Y
#> 31             <NA>    mm 122.78669     Y           NA          NA       Y
#> 32 107.698849780089    mm 122.78669        -15.0878390 -12.2878458       Y
#> 33 107.698849780089    mm 122.78669        -15.0878390 -12.2878458       Y
#> 34 181.575767859229    mm 122.78669         58.7890791  47.8790329       Y
#> 35 148.388321219484    mm 122.78669                 NA          NA        
#> 36 133.998835723584    mm  99.12705         34.8717828  35.1788758       Y
#> 37 204.663675562607    mm  99.12705        105.5366227 106.4660147       Y
#> 38 83.7624066358607    mm  99.12705        -15.3646463 -15.4999527       Y
#> 39             <NA>    mm  99.12705     Y           NA          NA       Y
#> 40 83.7624066358607    mm  99.12705        -15.3646463 -15.4999527       Y
#> 41 108.831916465708    mm  99.12705          9.7048636   9.7903280       Y
#> 42 179.879537693333    mm  99.12705                 NA          NA        
#> 43 138.391511890652    mm 141.45956         -3.0680515  -2.1688541       Y
#> 44 200.434648673004    mm 141.45956         58.9750853  41.6904194       Y
#> 45             <NA>    mm 141.45956     Y           NA          NA       Y
#> 46 198.011725546186    mm 141.45956         56.5521621  39.9776168       Y
#> 47 130.398197675437    mm 141.45956                 NA          NA        
#> 48 138.391511890652    mm 141.45956         -3.0680515  -2.1688541       Y
#> 49 161.600849239382    mm 141.45956         20.1412858  14.2381931       Y
#> 50 114.491808353182    mm 109.24628                 NA          NA        
#> 51  104.61987615837    mm 109.24628         -4.6264062  -4.2348409       Y
#> 52 112.406853018924    mm 109.24628          3.1605706   2.8930693       Y
#> 53             <NA>    mm 109.24628     Y           NA          NA       Y
#> 54  104.61987615837    mm 109.24628         -4.6264062  -4.2348409       Y
#> 55 150.229376163983    mm 109.24628         40.9830938  37.5144059       Y
#> 56 208.780712314368    mm 109.24628         99.5344299  91.1101300       Y
#> 57             <NA>    mm 131.96520     Y           NA          NA       Y
#> 58 124.721544072311    mm 131.96520                 NA          NA        
#> 59 140.571840594842    mm 131.96520          8.6066437   6.5219042       Y
#> 60 127.511097135508    mm 131.96520         -4.4540997  -3.3752079       Y
#> 61 127.511097135508    mm 131.96520         -4.4540997  -3.3752079       Y
#> 62 182.233782192385    mm 131.96520         50.2685853  38.0923051       Y
#> 63 157.817935052765    mm 131.96520         25.8527382  19.5905730       Y
#> 64 137.358792827839    mm 211.44121        -74.0824163 -35.0368864       Y
#> 65 139.474967305564    mm 211.44121        -71.9662418 -34.0360529       Y
#> 66 210.260613477461    mm 211.44121         -1.1805956  -0.5583564       Y
#> 67 137.358792827839    mm 211.44121        -74.0824163 -35.0368864       Y
#> 68 124.134050109426    mm 211.44121                 NA          NA        
#> 69 178.197602328401    mm 211.44121        -33.2436068 -15.7223878       Y
#> 70             <NA>    mm 211.44121     Y           NA          NA       Y
#>    ANL01FL ANL02FL ANL03FL
#> 1        Y       Y        
#> 2                        Y
#> 3                         
#> 4                Y       Y
#> 5        Y                
#> 6        Y                
#> 7        Y                
#> 8        Y                
#> 9        Y                
#> 10       Y       Y        
#> 11                       Y
#> 12               Y       Y
#> 13       Y                
#> 14                        
#> 15       Y       Y        
#> 16                       Y
#> 17                        
#> 18       Y                
#> 19       Y                
#> 20               Y       Y
#> 21       Y                
#> 22               Y       Y
#> 23                        
#> 24       Y                
#> 25       Y                
#> 26       Y                
#> 27                       Y
#> 28       Y       Y        
#> 29       Y                
#> 30       Y                
#> 31               Y       Y
#> 32       Y       Y        
#> 33                       Y
#> 34       Y                
#> 35                        
#> 36       Y                
#> 37       Y       Y        
#> 38       Y                
#> 39               Y       Y
#> 40                       Y
#> 41       Y                
#> 42                        
#> 43                       Y
#> 44       Y                
#> 45               Y       Y
#> 46       Y                
#> 47                        
#> 48       Y                
#> 49       Y       Y        
#> 50                        
#> 51       Y                
#> 52       Y                
#> 53               Y       Y
#> 54                       Y
#> 55       Y       Y        
#> 56       Y                
#> 57               Y       Y
#> 58                        
#> 59       Y                
#> 60       Y       Y        
#> 61                       Y
#> 62       Y                
#> 63       Y                
#> 64                       Y
#> 65       Y       Y        
#> 66       Y                
#> 67       Y                
#> 68                        
#> 69       Y                
#> 70               Y       Y