Adverse Event Analysis Dataset (ADAE)

Function for generating random Adverse Event Analysis Dataset for a given Subject-Level Analysis Dataset.

radae(ADSL, max_n_aes = 10, lookup = NULL, seed = NULL)

Arguments

ADSL
Subject-Level Analysis Dataset (ADSL).

max_n_aes
Maximum number of AEs per patient.

lookup
Data.frame containing additional parameters.

seed
Seed for random number generation.

Value

data.frame

Details

One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID AESTDTM AETERM AESEQ.

Examples

ADSL <- radsl(seed = 1)
ADAE <- radae(ADSL, seed = 2)
head(ADAE)

#> # A tibble: 6 x 21
#>   STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY ARM   ARMCD ACTARM
#>   <chr>   <chr>   <chr>  <chr>  <dbl> <fct> <fct> <fct>   <fct> <fct> <fct> 
#> 1 AB12345 AB1234… id-1   CHN-14    61 F     ASIAN CHN     C: C… ARM C C: Co…
#> 2 AB12345 AB1234… id-1   CHN-14    61 F     ASIAN CHN     C: C… ARM C C: Co…
#> 3 AB12345 AB1234… id-2   CHN-11    77 F     ASIAN CHN     C: C… ARM C C: Co…
#> 4 AB12345 AB1234… id-3   USA-17    27 F     AMER… USA     C: C… ARM C C: Co…
#> 5 AB12345 AB1234… id-3   USA-17    27 F     AMER… USA     C: C… ARM C C: Co…
#> 6 AB12345 AB1234… id-3   USA-17    27 F     AMER… USA     C: C… ARM C C: Co…
#> # … with 10 more variables: ACTARMCD <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <chr>, BEP01FL <fct>, AEBODSYS <chr>,
#> #   AEDECOD <chr>, AETOXGR <dbl>