Protocol Deviations Analysis Dataset (ADDV)

Function for generating random Protocol Deviations Analysis Dataset for a given Subject-Level Analysis Dataset.

raddv(
  ADSL,
  max_n_dv = 3L,
  p_dv = 0.15,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(ASTDT = c(seed = 1234, percentage = 0.1), DVCAT = c(seed = 1234,
    percentage = 0.1)),
  cached = FALSE
)

Arguments

ADSL
Subject-Level Analysis Dataset (ADSL).

max_n_dv
Maximum number of deviations per patient.

p_dv
Probability of a patient to have protocol deviations.

lookup
Data.frame containing additional parameters.

seed
Seed for random number generation.

na_percentage
(numeric) Default percentage of values to be replaced by NA

na_vars
(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached
boolean whether the cached ADDV data caddv should be returned or new data should be generated. If set to TRUE then the other arguments to raddv will be ignored.

Value

data.frame

Details

One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID ASTDT DVTERM DVSEQ.

Examples

ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
raddv(ADSL, seed = 2)

#> # A tibble: 9 x 49
#>   STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY DTHFL INVID INVNAM
#>   <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <fct> <chr> <chr> 
#> 1 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     N     INV … Dr. B…
#> 2 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     N     INV … Dr. B…
#> 3 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     N     INV … Dr. B…
#> 4 AB12345 AB1234… id-5   CHN-3     36 F     ASIAN CHN     N     INV … Dr. C…
#> 5 AB12345 AB1234… id-5   CHN-3     36 F     ASIAN CHN     N     INV … Dr. C…
#> 6 AB12345 AB1234… id-5   CHN-3     36 F     ASIAN CHN     N     INV … Dr. C…
#> 7 AB12345 AB1234… id-3   USA-13    35 F     AMER… USA     N     INV … Dr. U…
#> 8 AB12345 AB1234… id-3   USA-13    35 F     AMER… USA     N     INV … Dr. U…
#> 9 AB12345 AB1234… id-3   USA-13    35 F     AMER… USA     N     INV … Dr. U…
#> # … with 38 more variables: ARM <fct>, ARMCD <fct>, ACTARM <fct>,
#> #   ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>,
#> #   EOSDY <int>, DCSREAS <fct>, DTHDT <date>, DTHCAUS <fct>, DTHCAT <fct>,
#> #   LDDTHELD <int>, LDDTHGR1 <fct>, LSTALVDT <date>, study_duration_secs <dbl>,
#> #   DOMAIN <chr>, DVSEQ <fct>, DVTERM <fct>, DVDECOD <fct>, DVCAT <fct>,
#> #   DVSCAT <fct>, DVREAS <fct>, DVEPRELI <fct>, ASTDT <date>, ASTDY <int>,
#> #   AEPRELFL <fct>