Hys Law Analysis Dataset (ADHY)

Function for generating random dataset from Hys Law Analysis Dataset for a given Subject-Level Analysis Dataset.

radhy(
  ADSL,
  param = c("TBILI <= 2 times ULN and ALT value category",
    "TBILI > 2 times ULN and AST value category",
    "TBILI > 2 times ULN and ALT value category",
    "TBILI <= 2 times ULN and AST value category",
    "TBILI > 2 times ULN and ALKPH <= 2 times ULN and ALT value category",
    "TBILI > 2 times ULN and ALKPH <= 2 times ULN and AST value category",
    "TBILI > 2 times ULN and ALKPH <= 5 times ULN and ALT value category",
    "TBILI > 2 times ULN and ALKPH <= 5 times ULN and AST value category",
    "TBILI <= 2 times ULN and two consecutive elevations of ALT in relation to ULN",     
    "TBILI > 2 times ULN and two consecutive elevations of AST in relation to ULN",
    "TBILI <= 2 times ULN and two consecutive elevations of AST in relation to ULN",
    "TBILI > 2 times ULN and two consecutive elevations of ALT in relation to ULN",
    "TBILI > 2 times ULN and two consecutive elevations of ALT in relation to Baseline",
    "TBILI <= 2 times ULN and two consecutive elevations of ALT in relation to Baseline",
    "TBILI > 2 times ULN and two consecutive elevations of AST in relation to Baseline", 
       
    "TBILI <= 2 times ULN and two consecutive elevations of AST in relation to Baseline",
    "ALT > 3 times ULN by Period", "AST > 3 times ULN by Period",
    "ALT or AST > 3 times ULN by Period", "ALT > 3 times Baseline by Period",
    "AST > 3 times Baseline by Period", "ALT or AST > 3 times Baseline by Period"),
  paramcd = c("BLAL", "BGAS", "BGAL", "BLAS", "BA2AL", "BA2AS", "BA5AL", "BA5AS",
    "BL2AL2CU", "BG2AS2CU", "BL2AS2CU", "BG2AL2CU", "BG2AL2CB", "BL2AL2CB", "BG2AS2CB",
    "BL2AS2CB", "ALTPULN", "ASTPULN", "ALTASTPU", "ALTPBASE", "ASTPBASE", "ALTASTPB"),
  seed = NULL,
  cached = FALSE
)

Arguments

ADSL
Subject-Level Analysis Dataset (ADSL).

param
As character string. list of parameter values.

paramcd
As character string. list of parameter code values.

seed
Seed for random number generation.

cached
boolean whether the cached ADHY data cadhy should be returned or new data should be generated. If set to TRUE then the other arguments to radhy will be ignored.

Value

data.frame

Details

One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, AVISITN, ADTM, SRCSEQ.

Author

wojciakw

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radhy(ADSL, seed = 2)

#> # A tibble: 500 × 55
#>    STUDYID USUBJID  SUBJID SITEID   AGE SEX   RACE    COUNTRY DTHFL INVID INVNAM
#>    <chr>   <chr>    <chr>  <chr>  <int> <fct> <fct>   <fct>   <fct> <chr> <chr> 
#>  1 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  2 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  3 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  4 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  5 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  6 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  7 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  8 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  9 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#> 10 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#> # … with 490 more rows, and 44 more variables: ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>,
#> #   BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>,
#> #   EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>,
#> #   DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>,
#> #   LSTALVDT <date>, study_duration_secs <dbl>, PARAM <fct>, PARAMCD <fct>, …