Laboratory Data Analysis Dataset (ADLB)

Function for generating random dataset from Laboratory Data Analysis Dataset for a given Subject-Level Analysis Dataset.

radlb(
  ADSL,
  lbcat = c("CHEMISTRY", "CHEMISTRY", "IMMUNOLOGY"),
  param = c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement",
    "Immunoglobulin A Measurement"),
  paramcd = c("ALT", "CRP", "IGA"),
  paramu = c("U/L", "mg/L", "g/L"),
  visit_format = "WEEK",
  n_assessments = 5L,
  n_days = 5L,
  max_n_lbs = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(LOQFL = c(NA, 0.1), ABLFL2 = c(1234, 0.1), ABLFL = c(1235, 0.1), BASE2
    = c(NA, 0.1), BASE = c(NA, 0.1), CHG2 = c(1235, 0.1), PCHG2 = c(1235, 0.1), CHG =
    c(1234, 0.1), PCHG = c(1234, 0.1)),
  cached = FALSE
)

Arguments

ADSL
Subject-Level Analysis Dataset (ADSL).

lbcat
As character vector of lb category values.

param
As character string. list of parameter values.

paramcd
As character string. list of parameter code values.

paramu
As character vector of parameter unit values.

visit_format
Type of visit either "WEEK" or "CYCLE".

n_assessments
Number of weeks or cycles.

n_days
Number of days within cycle.

max_n_lbs
As numeric. maximum number of labs.

lookup
control lookup process.

seed
Seed for random number generation.

na_percentage
(numeric) Default percentage of values to be replaced by NA

na_vars
(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached
boolean whether the cached ADLB data cadlb should be returned or new data should be generated. If set to TRUE then the other arguments to radlb will be ignored.

Value

data.frame

Details

One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, BASETYPE, AVISITN, ATPTN, DTYPE, ADTM, LBSEQ, ASPID.

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radlb(ADSL, visit_format = "WEEK", n_assessments = 7L, seed = 2)

#> # A tibble: 270 x 59
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  2 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  3 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  4 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  5 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  6 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  7 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  8 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  9 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#> 10 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#> # … with 260 more rows, and 48 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>, ASEQ <int>,
#> #   LBSEQ <int>, LBTESTCD <fct>, LBTEST <fct>, LBCAT <fct>, LBSTRESC <chr>,
#> #   ASPID <int>, PARAM <fct>, PARAMCD <fct>, AVAL <dbl>, AVALU <fct>,
#> #   BASE2 <dbl>, BASE <dbl>, BASETYPE <fct>, ABLFL2 <fct>, ABLFL <fct>,
#> #   CHG2 <dbl>, PCHG2 <dbl>, CHG <dbl>, PCHG <dbl>, DTYPE <lgl>, ANRIND <fct>,
#> #   ADTM <dttm>, ADY <int>, ATPTN <int>, AVISIT <fct>, AVISITN <int>,
#> #   LOQFL <fct>

radlb(ADSL, visit_format = "CYCLE", n_assessments = 2L, seed = 2)

#> # A tibble: 330 x 59
#>    STUDYID USUBJID SUBJID SITEID   AGE SEX   RACE  COUNTRY INVID ARM   ARMCD
#>    <chr>   <chr>   <chr>  <chr>  <int> <fct> <fct> <fct>   <chr> <fct> <fct>
#>  1 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  2 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  3 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  4 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  5 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  6 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  7 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  8 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#>  9 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#> 10 AB12345 AB1234… id-9   BRA-1     35 F     BLAC… BRA     BRA-1 C: C… ARM C
#> # … with 320 more rows, and 48 more variables: ACTARM <fct>, ACTARMCD <fct>,
#> #   STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>,
#> #   SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>,
#> #   TRTEDTM <dttm>, EOSSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>,
#> #   DTHDT <date>, LSTALVDT <date>, study_duration_secs <dbl>, ASEQ <int>,
#> #   LBSEQ <int>, LBTESTCD <fct>, LBTEST <fct>, LBCAT <fct>, LBSTRESC <chr>,
#> #   ASPID <int>, PARAM <fct>, PARAMCD <fct>, AVAL <dbl>, AVALU <fct>,
#> #   BASE2 <dbl>, BASE <dbl>, BASETYPE <fct>, ABLFL2 <fct>, ABLFL <fct>,
#> #   CHG2 <dbl>, PCHG2 <dbl>, CHG <dbl>, PCHG <dbl>, DTYPE <lgl>, ANRIND <fct>,
#> #   ADTM <dttm>, ADY <int>, ATPTN <int>, AVISIT <fct>, AVISITN <int>,
#> #   LOQFL <fct>