Medical History Analysis Dataset (ADMH)

Function for generating random Medical History Analysis Dataset for a given Subject-Level Analysis Dataset.

radmh(
  ADSL,
  max_n_mhs = 10L,
  lookup = NULL,
  seed = NULL,
  na_percentage = 0,
  na_vars = list(MHBODSYS = c(NA, 0.1), MHDECOD = c(1234, 0.1)),
  cached = FALSE
)

Arguments

ADSL
Subject-Level Analysis Dataset (ADSL).

max_n_mhs
Maximum number of MHs per patient.

lookup
Data.frame containing additional parameters.

seed
Seed for random number generation.

na_percentage
(numeric) Default percentage of values to be replaced by NA

na_vars
(list) A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements

seed The seed to be used for this element - can be left NA

percentage How many element should be replaced. 0 is 0 % 1 is 100 %, can be left NA and default percentage is used. This will overwrite default percentage (percentage argument))

cached
boolean whether the cached ADMH data cadmh should be returned or new data should be generated. If set to TRUE then the other arguments to radmh will be ignored.

Value

data.frame

Details

One record per each record in the corresponding SDTM domain.

Keys: STUDYID USUBJID ASTDTM MHSEQ.

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, study_duration = 2, seed = 1)
radmh(ADSL, seed = 2)

#> # A tibble: 51 × 49
#>    STUDYID USUBJID  SUBJID SITEID   AGE SEX   RACE    COUNTRY DTHFL INVID INVNAM
#>    <chr>   <chr>    <chr>  <chr>  <int> <fct> <fct>   <fct>   <fct> <chr> <chr> 
#>  1 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  2 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  3 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  4 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  5 AB12345 AB12345… id-9   BRA-1     35 F     BLACK … BRA     N     INV … Dr. B…
#>  6 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#>  7 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#>  8 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#>  9 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#> 10 AB12345 AB12345… id-8   BRA-9     31 F     ASIAN   BRA     N     INV … Dr. B…
#> # … with 41 more rows, and 38 more variables: ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>,
#> #   BMRKR1 <dbl>, BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>,
#> #   BEP01FL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>, EOSSTT <fct>,
#> #   EOTSTT <fct>, EOSDT <date>, EOSDY <int>, DCSREAS <fct>, DTHDT <date>,
#> #   DTHCAUS <fct>, DTHCAT <fct>, LDDTHELD <int>, LDDTHGR1 <fct>,
#> #   LSTALVDT <date>, study_duration_secs <dbl>, ASEQ <int>, MHSEQ <int>, …