Subject-Level Analysis Dataset (ADSL)

The Subject-Level Analysis Dataset (ADSL) is used to provide the variables that describe attributes of a subject. ADSL is a source for subject-level variables used in other analysis data sets, such as population flags and treatment variables. There is only one ADSL per study. ADSL and its related metadata are required in a CDISC-based submission of data from a clinical trial even if no other analysis data sets are submitted.

Usage

radsl(
  N = 400,
  study_duration = 2,
  seed = NULL,
  with_trt02 = TRUE,
  na_percentage = 0,
  na_vars = list(AGE = NA, SEX = NA, RACE = NA, STRATA1 = NA, STRATA2 = NA, BMRKR1 =
    c(seed = 1234, percentage = 0.1), BMRKR2 = c(1234, 0.1), BEP01FL = NA),
  ae_withdrawal_prob = 0.05,
  cached = FALSE
)

Arguments

N

(numeric)
Number of patients.

study_duration

(numeric)
Duration of study in years.

seed

(numeric)
Seed to use for reproducible random number generation.

with_trt02

(logical)
Should period 2 be added.

na_percentage

(proportion)
Default percentage of values to be replaced by NA.

na_vars

(list)
A named list where the name of each element is a column name of ds. Each element of this list should be a numeric vector with two elements:

seed (numeric)
The seed to be used for this element - can be NA.

percentage (proportion)
Percentage of elements to be replaced with NA. If NA, na_percentage is used as a default.

ae_withdrawal_prob

(proportion)
Probability that there is at least one Adverse Event leading to the withdrawal of a study drug.

cached

boolean whether the cached ADSL data cadsl should be returned or new data should be generated. If set to TRUE then the other arguments to radsl will be ignored.

Value

data.frame

Details

One record per subject.

Keys: STUDYID, USUBJID

Examples

adsl <- radsl(N = 10, study_duration = 2, seed = 1)
adsl
#> # A tibble: 10 × 55
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  2 AB12345 AB12345-J… id-7   JPN-4     30 YEARS F     ASIAN NOT H… JPN     N    
#>  3 AB12345 AB12345-U… id-3   USA-13    35 YEARS F     AMER… NOT H… USA     N    
#>  4 AB12345 AB12345-B… id-8   BRA-9     31 YEARS F     ASIAN NOT H… BRA     Y    
#>  5 AB12345 AB12345-C… id-2   CHN-11    35 YEARS M     BLAC… NOT H… CHN     N    
#>  6 AB12345 AB12345-C… id-1   CHN-11    35 YEARS F     WHITE NOT H… CHN     Y    
#>  7 AB12345 AB12345-C… id-5   CHN-3     36 YEARS F     ASIAN NOT H… CHN     N    
#>  8 AB12345 AB12345-R… id-4   RUS-1     36 YEARS M     BLAC… NOT H… RUS     N    
#>  9 AB12345 AB12345-R… id-6   RUS-1     36 YEARS F     ASIAN NOT H… RUS     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> # ℹ 44 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>,
#> #   AP01SDTM <dttm>, AP01EDTM <dttm>, AP02SDTM <dttm>, AP02EDTM <dttm>, …

adsl <- radsl(
  N = 10, seed = 1,
  na_percentage = 0.1,
  na_vars = list(
    DTHDT = c(seed = 1234, percentage = 0.1),
    LSTALVDT = c(seed = 1234, percentage = 0.1)
  )
)
adsl
#> # A data frame: 10 × 55
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-C… id-10  CHN-3     35 YEARS M     ASIAN NOT H… CHN     N    
#>  2 AB12345 AB12345-J… id-7   JPN-4     30 YEARS F     ASIAN NOT H… JPN     N    
#>  3 AB12345 AB12345-U… id-3   USA-13    35 YEARS F     AMER… NOT H… USA     N    
#>  4 AB12345 AB12345-B… id-8   BRA-9     31 YEARS F     ASIAN NOT H… BRA     Y    
#>  5 AB12345 AB12345-C… id-2   CHN-11    35 YEARS M     BLAC… NOT H… CHN     N    
#>  6 AB12345 AB12345-C… id-1   CHN-11    35 YEARS F     WHITE NOT H… CHN     Y    
#>  7 AB12345 AB12345-C… id-5   CHN-3     36 YEARS F     ASIAN NOT H… CHN     N    
#>  8 AB12345 AB12345-R… id-4   RUS-1     36 YEARS M     BLAC… NOT H… RUS     N    
#>  9 AB12345 AB12345-R… id-6   RUS-1     36 YEARS F     ASIAN NOT H… RUS     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> # ℹ 44 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>,
#> #   AP01SDTM <dttm>, AP01EDTM <dttm>, AP02SDTM <dttm>, AP02EDTM <dttm>, …

adsl <- radsl(N = 10, seed = 1, na_percentage = .1)
adsl
#> # A data frame: 10 × 55
#>    STUDYID USUBJID    SUBJID SITEID   AGE AGEU  SEX   RACE  ETHNIC COUNTRY DTHFL
#>    <chr>   <chr>      <chr>  <chr>  <int> <fct> <fct> <fct> <fct>  <fct>   <fct>
#>  1 AB12345 AB12345-C… id-10  CHN-3     NA YEARS M     NA    NOT H… CHN     N    
#>  2 AB12345 AB12345-J… id-7   JPN-4     30 YEARS NA    ASIAN NOT H… JPN     N    
#>  3 AB12345 AB12345-U… id-3   USA-13    35 YEARS F     AMER… NOT H… USA     N    
#>  4 AB12345 AB12345-B… id-8   BRA-9     31 YEARS F     ASIAN NOT H… BRA     Y    
#>  5 AB12345 AB12345-C… id-2   CHN-11    35 YEARS M     BLAC… NOT H… CHN     N    
#>  6 AB12345 AB12345-C… id-1   CHN-11    35 YEARS F     WHITE NOT H… CHN     Y    
#>  7 AB12345 AB12345-C… id-5   CHN-3     36 YEARS F     ASIAN NOT H… CHN     N    
#>  8 AB12345 AB12345-R… id-4   RUS-1     36 YEARS M     BLAC… NOT H… RUS     N    
#>  9 AB12345 AB12345-R… id-6   RUS-1     36 YEARS F     ASIAN NOT H… RUS     N    
#> 10 AB12345 AB12345-B… id-9   BRA-1     35 YEARS F     BLAC… UNKNO… BRA     N    
#> # ℹ 44 more variables: INVID <chr>, INVNAM <chr>, ARM <fct>, ARMCD <fct>,
#> #   ACTARM <fct>, ACTARMCD <fct>, TRT01P <fct>, TRT01A <fct>, TRT02P <fct>,
#> #   TRT02A <fct>, REGION1 <fct>, STRATA1 <fct>, STRATA2 <fct>, BMRKR1 <dbl>,
#> #   BMRKR2 <fct>, ITTFL <fct>, SAFFL <fct>, BMEASIFL <fct>, BEP01FL <fct>,
#> #   AEWITHFL <fct>, RANDDT <date>, TRTSDTM <dttm>, TRTEDTM <dttm>,
#> #   TRT01SDTM <dttm>, TRT01EDTM <dttm>, TRT02SDTM <dttm>, TRT02EDTM <dttm>,
#> #   AP01SDTM <dttm>, AP01EDTM <dttm>, AP02SDTM <dttm>, AP02EDTM <dttm>, …