Tumor Response Analysis Dataset (ADTR)

Function for generating random dataset from Tumor Response Analysis Dataset for a given Subject-Level Analysis Dataset.

radtr(
  ADSL,
  param = c("Sum of Longest Diameter by Investigator"),
  paramcd = c("SLDINV"),
  seed = NULL,
  cached = FALSE,
  ...
)

Arguments

ADSL
Subject-Level Analysis Dataset (ADSL).

param
As character string. list of parameter values.

paramcd
As character string. list of parameter code values.

seed
Seed for random number generation.

cached
boolean whether the cached ADTR data cadtr should be returned or new data should be generated. If set to TRUE then the other arguments to radtr will be ignored.

...
parameters from radrs

Value

data.frame

Details

One record per subject per parameter per analysis visit per analysis date.

Keys: STUDYID, USUBJID, PARAMCD, BASETYPE, AVISITN, DTYPE

Examples

library(random.cdisc.data)
ADSL <- radsl(N = 10, seed = 1, study_duration = 2)
radtr(ADSL, seed = 2)

#>    STUDYID             USUBJID SUBJID SITEID AGE SEX
#> 1  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 2  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 3  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 4  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 5  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 6  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 7  AB12345  AB12345-BRA-1-id-9   id-9  BRA-1  35   F
#> 8  AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 9  AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 10 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 11 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 12 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 13 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 14 AB12345  AB12345-BRA-9-id-8   id-8  BRA-9  31   F
#> 15 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 16 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 17 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 18 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 19 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 20 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 21 AB12345 AB12345-CHN-11-id-1   id-1 CHN-11  35   F
#> 22 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 23 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 24 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 25 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 26 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 27 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 28 AB12345 AB12345-CHN-11-id-2   id-2 CHN-11  35   M
#> 29 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 30 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 31 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 32 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 33 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 34 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 35 AB12345 AB12345-CHN-3-id-10  id-10  CHN-3  35   M
#> 36 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 37 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 38 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 39 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 40 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 41 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 42 AB12345  AB12345-CHN-3-id-5   id-5  CHN-3  36   F
#> 43 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 44 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 45 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 46 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 47 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 48 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 49 AB12345  AB12345-JPN-4-id-7   id-7  JPN-4  30   F
#> 50 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 51 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 52 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 53 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 54 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 55 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 56 AB12345  AB12345-RUS-1-id-4   id-4  RUS-1  36   M
#> 57 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 58 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 59 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 60 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 61 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 62 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 63 AB12345  AB12345-RUS-1-id-6   id-6  RUS-1  36   F
#> 64 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 65 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 66 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 67 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 68 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 69 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#> 70 AB12345 AB12345-USA-13-id-3   id-3 USA-13  35   F
#>                                RACE COUNTRY DTHFL         INVID         INVNAM
#> 1         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 2         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 3         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 4         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 5         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 6         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 7         BLACK OR AFRICAN AMERICAN     BRA     N  INV ID BRA-1  Dr. BRA-1 Doe
#> 8                             ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 9                             ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 10                            ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 11                            ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 12                            ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 13                            ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 14                            ASIAN     BRA     N  INV ID BRA-9  Dr. BRA-9 Doe
#> 15                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 16                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 17                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 18                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 19                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 20                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 21                            WHITE     CHN     N INV ID CHN-11 Dr. CHN-11 Doe
#> 22        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 23        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 24        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 25        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 26        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 27        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 28        BLACK OR AFRICAN AMERICAN     CHN     Y INV ID CHN-11 Dr. CHN-11 Doe
#> 29                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 30                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 31                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 32                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 33                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 34                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 35                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 36                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 37                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 38                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 39                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 40                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 41                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 42                            ASIAN     CHN     N  INV ID CHN-3  Dr. CHN-3 Doe
#> 43                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 44                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 45                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 46                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 47                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 48                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 49                            ASIAN     JPN     N  INV ID JPN-4  Dr. JPN-4 Doe
#> 50        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 51        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 52        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 53        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 54        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 55        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 56        BLACK OR AFRICAN AMERICAN     RUS     N  INV ID RUS-1  Dr. RUS-1 Doe
#> 57                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 58                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 59                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 60                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 61                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 62                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 63                            ASIAN     RUS     Y  INV ID RUS-1  Dr. RUS-1 Doe
#> 64 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#> 65 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#> 66 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#> 67 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#> 68 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#> 69 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#> 70 AMERICAN INDIAN OR ALASKA NATIVE     USA     N INV ID USA-13 Dr. USA-13 Doe
#>               ARM ARMCD         ACTARM ACTARMCD       REGION1 STRATA1 STRATA2
#> 1  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 2  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 3  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 4  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 5  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 6  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 7  C: Combination ARM C C: Combination    ARM C South America       A      S1
#> 8       A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 9       A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 10      A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 11      A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 12      A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 13      A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 14      A: Drug X ARM A      A: Drug X    ARM A South America       B      S2
#> 15 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 16 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 17 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 18 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 19 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 20 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 21 C: Combination ARM C C: Combination    ARM C          Asia       B      S1
#> 22      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 23      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 24      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 25      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 26      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 27      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 28      A: Drug X ARM A      A: Drug X    ARM A          Asia       C      S2
#> 29     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 30     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 31     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 32     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 33     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 34     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 35     B: Placebo ARM B     B: Placebo    ARM B          Asia       A      S1
#> 36      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 37      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 38      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 39      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 40      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 41      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 42      A: Drug X ARM A      A: Drug X    ARM A          Asia       A      S2
#> 43      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 44      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 45      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 46      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 47      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 48      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 49      A: Drug X ARM A      A: Drug X    ARM A          Asia       B      S1
#> 50     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 51     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 52     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 53     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 54     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 55     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 56     B: Placebo ARM B     B: Placebo    ARM B       Eurasia       A      S1
#> 57      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 58      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 59      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 60      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 61      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 62      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 63      A: Drug X ARM A      A: Drug X    ARM A       Eurasia       B      S1
#> 64      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#> 65      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#> 66      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#> 67      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#> 68      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#> 69      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#> 70      A: Drug X ARM A      A: Drug X    ARM A North America       C      S1
#>       BMRKR1 BMRKR2 ITTFL SAFFL BMEASIFL BEP01FL     RANDDT             TRTSDTM
#> 1   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 2   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 3   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 4   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 5   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 6   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 7   6.922043   HIGH     Y     Y        Y       N 2021-01-17 2021-01-17 22:43:53
#> 8  11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 9  11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 10 11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 11 11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 12 11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 13 11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 14 11.277028 MEDIUM     Y     Y        Y       Y 2020-01-18 2020-01-21 10:19:08
#> 15  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 16  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 17  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 18  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 19  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 20  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 21  8.926907    LOW     Y     Y        N       N 2020-04-29 2020-05-02 02:28:45
#> 22  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 23  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 24  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 25  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 26  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 27  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 28  4.084350   HIGH     Y     Y        Y       Y 2020-02-23 2020-02-26 18:15:59
#> 29  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 30  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 31  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 32  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 33  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 34  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 35  1.777905    LOW     Y     Y        N       Y 2019-03-24 2019-03-28 12:07:16
#> 36  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 37  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 38  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 39  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 40  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 41  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 42  2.619796    LOW     Y     Y        Y       N 2020-06-24 2020-06-27 04:48:00
#> 43  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 44  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 45  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 46  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 47  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 48  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 49  7.459420 MEDIUM     Y     Y        Y       N 2019-09-22 2019-09-22 02:25:18
#> 50  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 51  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 52  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 53  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 54  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 55  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 56  9.952892 MEDIUM     Y     Y        Y       N 2020-08-27 2020-08-28 20:42:09
#> 57  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 58  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 59  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 60  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 61  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 62  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 63  9.342782 MEDIUM     Y     Y        N       N 2020-10-24 2020-10-26 22:16:26
#> 64  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#> 65  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#> 66  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#> 67  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#> 68  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#> 69  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#> 70  7.254058    LOW     Y     Y        Y       N 2019-10-16 2019-10-17 17:56:02
#>                TRTEDTM       EOSSTT      EOSDT EOSDY       DCSREAS      DTHDT
#> 1  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 2  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 3  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 4  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 5  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 6  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 7  2023-01-18 10:22:17    COMPLETED 2023-01-18   731          <NA>       <NA>
#> 8                 <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 9                 <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 10                <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 11                <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 12                <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 13                <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 14                <NA>      ONGOING       <NA>    NA          <NA>       <NA>
#> 15 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 16 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 17 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 18 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 19 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 20 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 21 2022-05-02 14:07:09    COMPLETED 2022-05-02   731          <NA>       <NA>
#> 22 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 23 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 24 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 25 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 26 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 27 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 28 2021-01-22 14:18:49 DISCONTINUED 2021-01-22   331         DEATH 2021-01-23
#> 29 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 30 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 31 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 32 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 33 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 34 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 35 2021-03-27 23:45:40    COMPLETED 2021-03-27   731          <NA>       <NA>
#> 36 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 37 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 38 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 39 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 40 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 41 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 42 2020-11-06 15:43:55 DISCONTINUED 2020-11-06   133 ADVERSE EVENT       <NA>
#> 43 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 44 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 45 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 46 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 47 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 48 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 49 2021-09-21 14:03:42    COMPLETED 2021-09-21   731          <NA>       <NA>
#> 50 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 51 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 52 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 53 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 54 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 55 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 56 2022-08-29 08:20:33    COMPLETED 2022-08-29   731          <NA>       <NA>
#> 57 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 58 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 59 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 60 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 61 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 62 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 63 2021-01-30 03:55:51 DISCONTINUED 2021-01-30    96         DEATH 2021-02-24
#> 64 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#> 65 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#> 66 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#> 67 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#> 68 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#> 69 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#> 70 2021-10-17 05:34:26    COMPLETED 2021-10-17   731          <NA>       <NA>
#>          DTHCAUS        DTHCAT LDDTHELD LDDTHGR1   LSTALVDT study_duration_secs
#> 1           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 2           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 3           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 4           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 5           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 6           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 7           <NA>          <NA>       NA     <NA> 2023-02-15            63113904
#> 8           <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 9           <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 10          <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 11          <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 12          <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 13          <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 14          <NA>          <NA>       NA     <NA>       <NA>            63113904
#> 15          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 16          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 17          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 18          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 19          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 20          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 21          <NA>          <NA>       NA     <NA> 2022-05-25            63113904
#> 22 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 23 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 24 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 25 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 26 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 27 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 28 ADVERSE EVENT ADVERSE EVENT        1     <=30 2021-01-23            63113904
#> 29          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 30          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 31          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 32          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 33          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 34          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 35          <NA>          <NA>       NA     <NA> 2021-04-10            63113904
#> 36          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 37          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 38          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 39          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 40          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 41          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 42          <NA>          <NA>       NA     <NA> 2020-11-17            63113904
#> 43          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 44          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 45          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 46          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 47          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 48          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 49          <NA>          <NA>       NA     <NA> 2021-10-03            63113904
#> 50          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 51          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 52          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 53          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 54          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 55          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 56          <NA>          <NA>       NA     <NA> 2022-09-10            63113904
#> 57 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 58 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 59 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 60 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 61 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 62 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 63 ADVERSE EVENT ADVERSE EVENT       25     <=30 2021-02-24            63113904
#> 64          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#> 65          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#> 66          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#> 67          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#> 68          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#> 69          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#> 70          <NA>          <NA>       NA     <NA> 2021-11-15            63113904
#>    AGEGR1   TRTDURD DCSREAS_GRP AEWITHFL
#> 1     <65 730.48499        <NA>     <NA>
#> 2     <65 730.48499        <NA>     <NA>
#> 3     <65 730.48499        <NA>     <NA>
#> 4     <65 730.48499        <NA>     <NA>
#> 5     <65 730.48499        <NA>     <NA>
#> 6     <65 730.48499        <NA>     <NA>
#> 7     <65 730.48499        <NA>     <NA>
#> 8     <65        NA        <NA>     <NA>
#> 9     <65        NA        <NA>     <NA>
#> 10    <65        NA        <NA>     <NA>
#> 11    <65        NA        <NA>     <NA>
#> 12    <65        NA        <NA>     <NA>
#> 13    <65        NA        <NA>     <NA>
#> 14    <65        NA        <NA>     <NA>
#> 15    <65 730.48499        <NA>     <NA>
#> 16    <65 730.48499        <NA>     <NA>
#> 17    <65 730.48499        <NA>     <NA>
#> 18    <65 730.48499        <NA>     <NA>
#> 19    <65 730.48499        <NA>     <NA>
#> 20    <65 730.48499        <NA>     <NA>
#> 21    <65 730.48499        <NA>     <NA>
#> 22    <65 330.83529  Non-Safety        N
#> 23    <65 330.83529  Non-Safety        N
#> 24    <65 330.83529  Non-Safety        N
#> 25    <65 330.83529  Non-Safety        N
#> 26    <65 330.83529  Non-Safety        N
#> 27    <65 330.83529  Non-Safety        N
#> 28    <65 330.83529  Non-Safety        N
#> 29    <65 730.48499        <NA>     <NA>
#> 30    <65 730.48499        <NA>     <NA>
#> 31    <65 730.48499        <NA>     <NA>
#> 32    <65 730.48499        <NA>     <NA>
#> 33    <65 730.48499        <NA>     <NA>
#> 34    <65 730.48499        <NA>     <NA>
#> 35    <65 730.48499        <NA>     <NA>
#> 36    <65 132.45549      Safety        Y
#> 37    <65 132.45549      Safety        Y
#> 38    <65 132.45549      Safety        Y
#> 39    <65 132.45549      Safety        Y
#> 40    <65 132.45549      Safety        Y
#> 41    <65 132.45549      Safety        Y
#> 42    <65 132.45549      Safety        Y
#> 43    <65 730.48499        <NA>     <NA>
#> 44    <65 730.48499        <NA>     <NA>
#> 45    <65 730.48499        <NA>     <NA>
#> 46    <65 730.48499        <NA>     <NA>
#> 47    <65 730.48499        <NA>     <NA>
#> 48    <65 730.48499        <NA>     <NA>
#> 49    <65 730.48499        <NA>     <NA>
#> 50    <65 730.48499        <NA>     <NA>
#> 51    <65 730.48499        <NA>     <NA>
#> 52    <65 730.48499        <NA>     <NA>
#> 53    <65 730.48499        <NA>     <NA>
#> 54    <65 730.48499        <NA>     <NA>
#> 55    <65 730.48499        <NA>     <NA>
#> 56    <65 730.48499        <NA>     <NA>
#> 57    <65  95.23569  Non-Safety        N
#> 58    <65  95.23569  Non-Safety        N
#> 59    <65  95.23569  Non-Safety        N
#> 60    <65  95.23569  Non-Safety        N
#> 61    <65  95.23569  Non-Safety        N
#> 62    <65  95.23569  Non-Safety        N
#> 63    <65  95.23569  Non-Safety        N
#> 64    <65 730.48499        <NA>     <NA>
#> 65    <65 730.48499        <NA>     <NA>
#> 66    <65 730.48499        <NA>     <NA>
#> 67    <65 730.48499        <NA>     <NA>
#> 68    <65 730.48499        <NA>     <NA>
#> 69    <65 730.48499        <NA>     <NA>
#> 70    <65 730.48499        <NA>     <NA>
#>                                      PARAM PARAMCD                AVISIT
#> 1  Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 2  Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 3  Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 4  Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 5  Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 6  Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 7  Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 8  Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 9  Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 10 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 11 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 12 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 13 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 14 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 15 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 16 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 17 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 18 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 19 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 20 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 21 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 22 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 23 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 24 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 25 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 26 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 27 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 28 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 29 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 30 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 31 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 32 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 33 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 34 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 35 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 36 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 37 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 38 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 39 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 40 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 41 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 42 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 43 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 44 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 45 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 46 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 47 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 48 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 49 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 50 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 51 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 52 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 53 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 54 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 55 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 56 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 57 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#> 58 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 59 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 60 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 61 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 62 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 63 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 64 Sum of Longest Diameter by Investigator  SLDINV POST-BASELINE MINIMUM
#> 65 Sum of Longest Diameter by Investigator  SLDINV      END OF INDUCTION
#> 66 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 4 DAY 1
#> 67 Sum of Longest Diameter by Investigator  SLDINV             FOLLOW UP
#> 68 Sum of Longest Diameter by Investigator  SLDINV             SCREENING
#> 69 Sum of Longest Diameter by Investigator  SLDINV         CYCLE 2 DAY 1
#> 70 Sum of Longest Diameter by Investigator  SLDINV              BASELINE
#>    AVISITN                ADTM  ADY   DTYPE      AVAL            AVALC AVALU
#> 1      999 2021-12-31 00:00:00  348          91.02365 91.0236527565802    mm
#> 2      999 2021-12-31 00:00:00  348 MINIMUM  91.02365 91.0236527565802    mm
#> 3       -1 2020-10-09 00:00:00 -100         124.85139 124.851385571939    mm
#> 4        0 2021-01-17 22:43:53    0                NA             <NA>    mm
#> 5        2 2021-08-08 00:00:00  203         133.13259 133.132588399725    mm
#> 6        4 2021-10-07 00:00:00  263         188.27147 188.271465355528    mm
#> 7      999 2021-10-16 00:00:00  272         118.57282 118.572821190151    mm
#> 8      999 2021-04-28 00:00:00  463          96.35273 96.3527337970544    mm
#> 9        4                <NA>   NA         200.14856  200.14856301978    mm
#> 10     999 2021-06-05 00:00:00  501         210.93728 210.937275562657    mm
#> 11     999 2021-04-28 00:00:00  463 MINIMUM  96.35273 96.3527337970544    mm
#> 12       0 2020-01-21 10:19:08    0                NA             <NA>    mm
#> 13       2 2021-05-14 00:00:00  479         184.17689 184.176894082022    mm
#> 14      -1 2019-10-13 00:00:00 -100         140.31087 140.310867187201    mm
#> 15       4 2022-03-16 00:00:00  683         125.40015 125.400146831514    mm
#> 16     999 2021-02-14 00:00:00  288 MINIMUM  90.03459 90.0345901510763    mm
#> 17      -1 2020-01-23 00:00:00 -100         128.90567 128.905670010183    mm
#> 18       2 2022-01-15 00:00:00  623         165.18704 165.187043924422    mm
#> 19     999 2021-02-14 00:00:00  288          90.03459 90.0345901510763    mm
#> 20       0 2020-05-02 02:28:45    0                NA             <NA>    mm
#> 21     999 2021-11-18 00:00:00  565         135.62122 135.621222256451    mm
#> 22       0 2020-02-26 18:15:59    0                NA             <NA>    mm
#> 23      -1 2019-11-18 00:00:00 -100         152.52540 152.525397129667    mm
#> 24     999 2020-07-18 00:00:00  143         163.04543 163.045432845081    mm
#> 25       2 2020-10-20 00:00:00  237         122.36173 122.361730034768    mm
#> 26     999 2020-04-10 00:00:00   44         145.75018 145.750175731918    mm
#> 27       2 2020-10-20 00:00:00  237 MINIMUM 122.36173 122.361730034768    mm
#> 28       4 2020-12-19 00:00:00  297         159.91349 159.913485103959    mm
#> 29       2 2020-03-11 00:00:00  349         189.10537 189.105366972668    mm
#> 30       4 2020-05-10 00:00:00  409         173.15369 173.153693266196    mm
#> 31       0 2019-03-28 12:07:16    0                NA             <NA>    mm
#> 32     999 2020-05-29 00:00:00  428         107.69885 107.698849780089    mm
#> 33     999 2020-05-29 00:00:00  428 MINIMUM 107.69885 107.698849780089    mm
#> 34     999 2020-04-20 00:00:00  389         181.57577 181.575767859229    mm
#> 35      -1 2018-12-18 00:00:00 -100         148.38832 148.388321219484    mm
#> 36       2 2020-08-06 00:00:00   40         133.99884 133.998835723584    mm
#> 37     999 2020-11-02 00:00:00  128         204.66368 204.663675562607    mm
#> 38     999 2020-08-13 00:00:00   47          83.76241 83.7624066358607    mm
#> 39       0 2020-06-27 04:48:00    0                NA             <NA>    mm
#> 40     999 2020-08-13 00:00:00   47 MINIMUM  83.76241 83.7624066358607    mm
#> 41       4 2020-10-05 00:00:00  100         108.83192 108.831916465708    mm
#> 42      -1 2020-03-19 00:00:00 -100         179.87954 179.879537693333    mm
#> 43       2 2021-07-13 00:00:00  660 MINIMUM 138.39151 138.391511890652    mm
#> 44     999 2021-03-23 00:00:00  548         200.43465 200.434648673004    mm
#> 45       0 2019-09-22 02:25:18    0                NA             <NA>    mm
#> 46     999 2020-05-04 00:00:00  225         198.01173 198.011725546186    mm
#> 47      -1 2019-06-14 00:00:00 -100         130.39820 130.398197675437    mm
#> 48       2 2021-07-13 00:00:00  660         138.39151 138.391511890652    mm
#> 49       4 2021-09-11 00:00:00  720         161.60085 161.600849239382    mm
#> 50      -1 2020-05-20 00:00:00 -100         114.49181 114.491808353182    mm
#> 51       2 2021-11-22 00:00:00  451         104.61988  104.61987615837    mm
#> 52       4 2022-01-21 00:00:00  511         112.40685 112.406853018924    mm
#> 53       0 2020-08-28 20:42:09    0                NA             <NA>    mm
#> 54       2 2021-11-22 00:00:00  451 MINIMUM 104.61988  104.61987615837    mm
#> 55     999 2022-06-26 00:00:00  667         150.22938 150.229376163983    mm
#> 56     999 2022-03-06 00:00:00  555         208.78071 208.780712314368    mm
#> 57       0 2020-10-26 22:16:26    0                NA             <NA>    mm
#> 58      -1 2020-07-18 00:00:00 -100         124.72154 124.721544072311    mm
#> 59     999 2020-12-15 00:00:00   50         140.57184 140.571840594842    mm
#> 60     999 2021-01-29 00:00:00   95         127.51110 127.511097135508    mm
#> 61     999 2021-01-29 00:00:00   95 MINIMUM 127.51110 127.511097135508    mm
#> 62       2 2020-11-09 00:00:00   14         182.23378 182.233782192385    mm
#> 63       4 2021-01-08 00:00:00   74         157.81794 157.817935052765    mm
#> 64     999 2020-09-13 00:00:00  332 MINIMUM 137.35879 137.358792827839    mm
#> 65     999 2020-11-08 00:00:00  388         139.47497 139.474967305564    mm
#> 66       4 2020-05-29 00:00:00  225         210.26061 210.260613477461    mm
#> 67     999 2020-09-13 00:00:00  332         137.35879 137.358792827839    mm
#> 68      -1 2019-07-09 00:00:00 -100         124.13405 124.134050109426    mm
#> 69       2 2020-03-30 00:00:00  165         178.19760 178.197602328401    mm
#> 70       0 2019-10-17 17:56:02    0                NA             <NA>    mm
#>         BASE ABLFL          CHG        PCHG ONTRTFL ANL01FL ANL02FL ANL03FL
#> 1  211.98904       -120.9653879 -57.0620951       Y       Y       Y        
#> 2  211.98904       -120.9653879 -57.0620951       Y                       Y
#> 3  211.98904                 NA          NA                                
#> 4  211.98904     Y           NA          NA       Y               Y       Y
#> 5  211.98904        -78.8564523 -37.1983627       Y       Y                
#> 6  211.98904        -23.7175753 -11.1881139       Y       Y                
#> 7  211.98904        -93.4162195 -44.0665325       Y       Y                
#> 8  178.07588        -81.7231420 -45.8923151       Y       Y                
#> 9  178.07588         22.0726872  12.3951024       Y       Y                
#> 10 178.07588         32.8613998  18.4535944       Y       Y       Y        
#> 11 178.07588        -81.7231420 -45.8923151       Y                       Y
#> 12 178.07588     Y           NA          NA       Y               Y       Y
#> 13 178.07588          6.1010183   3.4260779       Y       Y                
#> 14 178.07588                 NA          NA                                
#> 15 154.74494        -29.3447960 -18.9633312       Y       Y       Y        
#> 16 154.74494        -64.7103527 -41.8174265       Y                       Y
#> 17 154.74494                 NA          NA                                
#> 18 154.74494         10.4421010   6.7479433       Y       Y                
#> 19 154.74494        -64.7103527 -41.8174265       Y       Y                
#> 20 154.74494     Y           NA          NA       Y               Y       Y
#> 21 154.74494        -19.1237206 -12.3582201       Y       Y                
#> 22 123.13540     Y           NA          NA       Y               Y       Y
#> 23 123.13540                 NA          NA                                
#> 24 123.13540         39.9100312  32.4115004       Y       Y                
#> 25 123.13540         -0.7736716  -0.6283097       Y       Y                
#> 26 123.13540         22.6147741  18.3657776       Y       Y                
#> 27 123.13540         -0.7736716  -0.6283097       Y                       Y
#> 28 123.13540         36.7780834  29.8680013       Y       Y       Y        
#> 29 122.78669         66.3186782  54.0112930       Y       Y                
#> 30 122.78669         50.3670045  41.0199225       Y       Y                
#> 31 122.78669     Y           NA          NA       Y               Y       Y
#> 32 122.78669        -15.0878390 -12.2878458       Y       Y       Y        
#> 33 122.78669        -15.0878390 -12.2878458       Y                       Y
#> 34 122.78669         58.7890791  47.8790329       Y       Y                
#> 35 122.78669                 NA          NA                                
#> 36  99.12705         34.8717828  35.1788758       Y       Y                
#> 37  99.12705        105.5366227 106.4660147       Y       Y       Y        
#> 38  99.12705        -15.3646463 -15.4999527       Y       Y                
#> 39  99.12705     Y           NA          NA       Y               Y       Y
#> 40  99.12705        -15.3646463 -15.4999527       Y                       Y
#> 41  99.12705          9.7048636   9.7903280       Y       Y                
#> 42  99.12705                 NA          NA                                
#> 43 141.45956         -3.0680515  -2.1688541       Y                       Y
#> 44 141.45956         58.9750853  41.6904194       Y       Y                
#> 45 141.45956     Y           NA          NA       Y               Y       Y
#> 46 141.45956         56.5521621  39.9776168       Y       Y                
#> 47 141.45956                 NA          NA                                
#> 48 141.45956         -3.0680515  -2.1688541       Y       Y                
#> 49 141.45956         20.1412858  14.2381931       Y       Y       Y        
#> 50 109.24628                 NA          NA                                
#> 51 109.24628         -4.6264062  -4.2348409       Y       Y                
#> 52 109.24628          3.1605706   2.8930693       Y       Y                
#> 53 109.24628     Y           NA          NA       Y               Y       Y
#> 54 109.24628         -4.6264062  -4.2348409       Y                       Y
#> 55 109.24628         40.9830938  37.5144059       Y       Y       Y        
#> 56 109.24628         99.5344299  91.1101300       Y       Y                
#> 57 131.96520     Y           NA          NA       Y               Y       Y
#> 58 131.96520                 NA          NA                                
#> 59 131.96520          8.6066437   6.5219042       Y       Y                
#> 60 131.96520         -4.4540997  -3.3752079       Y       Y       Y        
#> 61 131.96520         -4.4540997  -3.3752079       Y                       Y
#> 62 131.96520         50.2685853  38.0923051       Y       Y                
#> 63 131.96520         25.8527382  19.5905730       Y       Y                
#> 64 211.44121        -74.0824163 -35.0368864       Y                       Y
#> 65 211.44121        -71.9662418 -34.0360529       Y       Y       Y        
#> 66 211.44121         -1.1805956  -0.5583564       Y       Y                
#> 67 211.44121        -74.0824163 -35.0368864       Y       Y                
#> 68 211.44121                 NA          NA                                
#> 69 211.44121        -33.2436068 -15.7223878       Y       Y                
#> 70 211.44121     Y           NA          NA       Y               Y       Y